重构tianyuan/按功能拆为4个域
