+++ custom_components/tianyuan_calendar/tianyuan/library/__init__.py
@@ -1,0 +1,18 @@
+"""天元资源库 - 加密资源加载与访问."""
+from .maps_loader import (
+    加载全量映射表类,
+    获取单卦基础信息类,
+    获取单卦详注类,
+    获取所有卦名列表类,
+    获取系统校验码法,
+    检查专业权限类,
+)
+
+__all__ = [
+    "加载全量映射表类",
+    "获取单卦基础信息类",
+    "获取单卦详注类",
+    "获取所有卦名列表类",
+    "获取系统校验码法",
+    "检查专业权限类",
+]
