--- custom_components/tianyuan_calendar/tianyuan/qihuang.py
@@ -1,1167 +1,0 @@
-"""天元岐黄引擎."""
-# 干支纪日：晚子时日柱算当天
-# 干支纪年：新年以立春节气交接的时刻起算
-from __future__ import annotations
-
-from .maps_loader import 加载全量映射表类
-
-maps = 加载全量映射表类()
-子午流注映射 = maps.get("子午流注类", {})
-五运六气映射 = maps.get("五运六气类", {})
-岐黄类映射 = maps.get("岐黄类", {})
-
-# 中医子午流注类 (ZiwuLiuzhu)
-class 子午流注类:
-    """子午流注：纳甲法 / 纳子法 / 灵龟八法 / 飞腾八法 / 迎随补泻"""
-
-    穴位手册 = 子午流注映射.get("穴位手册", {})
-    时辰经络 = 子午流注映射.get("时辰经络", [])
-    流注顺序 = 子午流注映射.get("十二经流注顺序", [])
-    子午相对 = 子午流注映射.get("子午相对", [])
-    纳甲取穴 = 子午流注映射.get("纳甲取穴", [])
-    值日经 = 子午流注映射.get("六十甲子值日经", [])
-    八卦方位 = 子午流注映射.get("八卦方位", {})
-    八卦吉凶 = 子午流注映射.get("八卦吉凶", {})
-    飞腾起点 = 子午流注映射.get("飞腾起点", "")
-    五输穴表 = 子午流注映射.get("五输穴表", {})
-    纳子补泻表 = 子午流注映射.get("纳子补泻表", {})
-    飞腾配穴表 = 子午流注映射.get("飞腾配穴表", {})
-
-    # ========= 公共工具 =========
-    @staticmethod
-    def 小时转时支(小时: int) -> str:
-        对照 = [
-            "子","丑","丑","寅","寅","卯","卯","辰","辰","巳","巳","午",
-            "午","未","未","申","申","酉","酉","戌","戌","亥","亥","子"
-        ]
-        return 对照[小时]
-
-    @staticmethod
-    def 穴位详情(穴位: str) -> dict:
-        return 子午流注类.穴位手册.get(穴位, {"位置": "未知", "功能": "未知", "针法": "未知"})
-
-    @staticmethod
-    def 生成描述(农历) -> str:
-        日干支 = 农历.getDayInGanZhiExact2()
-        时干支 = 农历.getTimeInGanZhi()
-        时支 = 农历.getTimeZhi()
-        return f"{日干支}日{时干支}时（{时支}时）"
-
-    @staticmethod
-    def 顺经(经络: str) -> str:
-        seq = 子午流注类.流注顺序
-        if 经络 not in seq:
-            return "未知"
-        return seq[(seq.index(经络) + 1) % len(seq)]
-
-    @staticmethod
-    def 逆经(经络: str) -> str:
-        seq = 子午流注类.流注顺序
-        if 经络 not in seq:
-            return "未知"
-        return seq[(seq.index(经络) - 1) % len(seq)]
-
-    @staticmethod
-    def 八卦信息(卦: str) -> tuple:
-        return (
-            子午流注类.八卦方位.get(卦, "未知"),
-            子午流注类.八卦吉凶.get(卦, "未知"),
-        )
-
-    @staticmethod
-    def 判开阖枢(时支: str) -> dict:
-        开 = ["子", "午", "卯", "酉"]
-        阖 = ["丑", "辰", "未", "戌"]
-        枢 = ["寅", "巳", "申", "亥"]
-
-        if 时支 in 开:
-            return {"法": "开", "说明": "开法：气血初开，多取井穴、荥穴，以开其端。"}
-        if 时支 in 阖:
-            return {"法": "阖", "说明": "阖法：气血闭藏，多取输穴、合穴，以收其气。"}
-        if 时支 in 枢:
-            return {"法": "枢", "说明": "枢法：气血转折，多取经穴、合穴，以调其枢机。"}
-        return {"法": "未知", "说明": "未能判定开/阖/枢。"}
-    
-    # ========= 纳甲法=========
-    @staticmethod
-    def 纳甲法类(农历):
-        """
-        严格按徐文伯纳甲法：阳日阳时开阳经，阴日阴时开阴经。
-        井荥输经合逐时顺移，闭穴时用合日互用（甲与己合等）。
-        返回当前时辰应开的穴位（含五行属性自动推算）。
-        """
-        日干 = 农历.getDayGanExact2()
-        时干 = 农历.getTimeGan()
-        时支 = 农历.getTimeZhi()
-        日干支 = 农历.getDayInGanZhiExact2()
-        当前时干支 = f"{时干}{时支}"
-
-        # 合日互用配对
-        合日 = {"甲":"己","己":"甲","乙":"庚","庚":"乙","丙":"辛","辛":"丙",
-                "丁":"壬","壬":"丁","戊":"癸","癸":"戊"}
-
-        # 阳经/阴经顺序（纳甲经典顺序）
-        阳经序 = ["胆经","小肠经","胃经","大肠经","膀胱经","三焦经"]
-        阴经序 = ["肝经","心经","脾经","肺经","肾经","心包经"]
-
-        # 开井时辰表（日干 -> (开井时辰干支, 经络, 井穴)）
-        开井表 = {
-            "甲": ("甲戌", "胆经", "足窍阴"),
-            "乙": ("乙酉", "肝经", "大敦"),
-            "丙": ("丙申", "小肠经", "少泽"),
-            "丁": ("丁未", "心经", "少冲"),
-            "戊": ("戊午", "胃经", "厉兑"),
-            "己": ("己巳", "脾经", "隐白"),
-            "庚": ("庚辰", "大肠经", "商阳"),
-            "辛": ("辛卯", "肺经", "少商"),
-            "壬": ("壬寅", "膀胱经", "至阴"),
-            "癸": ("癸亥", "肾经", "涌泉")
-        }
-
-        # 五输穴的五行属性（井荥输经合）
-        # 阳经：井金 荥水 输木 经火 合土
-        # 阴经：井木 荥火 输土 经金 合水
-        阳经五输五行 = {"井":"金", "荥":"水", "输":"木", "经":"火", "合":"土"}
-        阴经五输五行 = {"井":"木", "荥":"火", "输":"土", "经":"金", "合":"水"}
-
-        # 工具函数：从开井时辰开始，遍历12个时辰，查找当前时辰对应的穴位
-        def 从开井时辰查找(开井干支, 经络, 井穴):
-            开井干 = 开井干支[0]
-            开井支 = 开井干支[1]
-            天干序 = ["甲","乙","丙","丁","戊","己","庚","辛","壬","癸"]
-            地支序 = ["子","丑","寅","卯","辰","巳","午","未","申","酉","戌","亥"]
-            天干始 = 天干序.index(开井干)
-            地支始 = 地支序.index(开井支)
-
-            # 确定经序
-            if 经络 in 阳经序:
-                经序 = 阳经序
-            else:
-                经序 = 阴经序
-            经索引 = 经序.index(经络)
-
-            输穴列表 = ["井","荥","输","经","合"]
-            穴序 = 0   # 从井穴开始
-
-            for 偏移 in range(12):
-                干 = 天干序[(天干始 + 偏移) % 10]
-                支 = 地支序[(地支始 + 偏移) % 12]
-                干支 = f"{干}{支}"
-                if 干支 == 当前时干支:
-                    当前经 = 经序[经索引]
-                    输穴类型 = 输穴列表[穴序]
-                    穴位 = 子午流注类.五输穴表[当前经][输穴类型]
-                    五行 = (阳经五输五行 if 当前经 in 阳经序 else 阴经五输五行)[输穴类型]
-                    return 穴位, 输穴类型, 当前经, 五行
-                # 移到下一个穴或经
-                穴序 += 1
-                if 穴序 >= 5:   # 井荥输经合走完
-                    穴序 = 0
-                    经索引 = (经索引 + 1) % len(经序)
-            return None, None, None, None
-
-        # 先试本日开井
-        穴位结果 = None
-        if 日干 in 开井表:
-            开井干支, 经络, 井穴 = 开井表[日干]
-            穴位结果, 输穴类型, 使用经络, 五行属性 = 从开井时辰查找(开井干支, 经络, 井穴)
-
-        # 若未找到，试合日开井
-        if not 穴位结果:
-            合日干 = 合日.get(日干)
-            if 合日干 and 合日干 in 开井表:
-                开井干支, 经络, 井穴 = 开井表[合日干]
-                穴位结果, 输穴类型, 使用经络, 五行属性 = 从开井时辰查找(开井干支, 经络, 井穴)
-
-        # 组装返回数据
-        描述 = 子午流注类.生成描述(农历)
-        值日经 = 子午流注类.值日经.get(日干支, "未知")
-        三法 = 子午流注类.判开阖枢(时支)
-
-        if 穴位结果:
-            穴位详情 = 子午流注类.穴位详情(穴位结果)
-            摘要 = f"{穴位结果}（{输穴类型}，{使用经络}，五行{五行属性}），采用{三法['法']}法"
-            return {
-                "state": 穴位结果,
-                "attributes": {
-                    "值日经": 值日经,
-                    "穴位": 穴位结果,
-                    "输穴": 输穴类型,
-                    "经络": 使用经络,
-                    "五行": 五行属性,
-                    "三法": 三法["法"],
-                    "三法说明": 三法["说明"],
-                    "位置": 穴位详情['位置'],
-                    "功能": 穴位详情['功能'],
-                    "针法": 穴位详情['针法'],
-                    "摘要": 摘要,
-                    "描述": 描述,
-                    "详细描述": (
-                        f"{描述}，值日经：{值日经}，采用{三法['法']}法，"
-                        f"取{使用经络}之{输穴类型}——{穴位结果}（五行{五行属性}）。\n"
-                        f"穴位位置：{穴位详情['位置']}\n"
-                        f"主要功能：{穴位详情['功能']}\n"
-                        f"针刺方法：{穴位详情['针法']}\n"
-                        f"三法说明：{三法['说明']}"
-                    )
-                }
-            }
-        else:
-            # 闭穴，但尽量给出合日穴位的提示（如果合日仍闭穴，则真的闭穴）
-            提示 = "闭穴，请用合日互用或纳子法"
-            合日干 = 合日.get(日干)
-            if 合日干 and 合日干 in 开井表:
-                提示 += f"（可尝试{合日干}日开穴）"
-            return {
-                "state": "闭穴",
-                "attributes": {
-                    "描述": 描述,
-                    "详细描述": f"{描述}，当前时辰无纳甲开穴。{提示}"
-                }
-            }
-
-    # ========= 纳子法（补泻增强版） =========
-    @staticmethod
-    def 纳子法类(dt, 证型: str = "平"):
-        """
-        证型：平（取本穴/原穴）、实（取子穴）、虚（取母穴，下时辰取）
-        """
-        小时 = dt.hour
-        时支 = 子午流注类.小时转时支(小时)
-        经络, 默认穴, 默认输穴, 时间段 = 子午流注类.时辰经络[时支]
-        补泻信息 = 子午流注类.纳子补泻表.get(时支, {})
-        穴位详情 = {}
-        穴位 = 默认穴
-        输穴类型 = 默认输穴
-
-        if 证型 == "实":
-            实证穴 = 补泻信息.get("实证穴", 默认穴)
-            穴位 = 实证穴
-            输穴类型 = "子穴"
-            补泻描述 = f"实证泻子：取{经络}子穴{穴位}"
-        elif 证型 == "虚":
-            虚证穴 = 补泻信息.get("虚证穴", 默认穴)
-            虚证时辰 = 补泻信息.get("虚证时辰", 时支)
-            穴位 = 虚证穴
-            输穴类型 = "母穴"
-            # 提示应在下一个时辰取穴，但为了方便，这里仍显示穴位，并在描述中提示
-            补泻描述 = f"虚证补母：宜在{虚证时辰}时取{经络}母穴{穴位}"
-        else:
-            补泻描述 = f"平补平泻：取本穴/原穴{穴位}"
-
-        穴位详情 = 子午流注类.穴位详情(穴位)
-        对应时支 = 子午流注类.子午相对.get(时支)
-        对应经络 = 子午流注类.时辰经络[对应时支][0] if 对应时支 else "未知"
-
-        辅行诀参考 = 辅行诀脏腑用药法要类.子午流注纳子法辅行诀选方(经络, 证型)
-        伤寒杂病论参考 = 伤寒杂病论类.子午流注纳子法伤寒选方(经络, 证型)
-
-        描述 = f"{时支}时（{时间段}）"
-        摘要 = f"{时支}时当令{经络}，{补泻描述}，子午相对：{对应经络}"
-
-        return {
-            "state": 经络,
-            "attributes": {
-                "时支": 时支,
-                "时间段": 时间段,
-                "经络": 经络,
-                "穴位": 穴位,
-                "输穴": 输穴类型,
-                "子午相对时支": 对应时支,
-                "子午相对经络": 对应经络,
-                "位置": 穴位详情['位置'],
-                "功能": 穴位详情['功能'],
-                "针法": 穴位详情['针法'],
-                "摘要": 摘要,
-                "描述": 描述,
-                "辅行诀参考": 辅行诀参考,
-                "经方参考": 伤寒杂病论参考,
-                "详细描述": (
-                    f"{描述}，气血流注于{经络}。\n"
-                    f"{补泻描述}。\n"
-                    f"子午相对：{时支} ↔ {对应时支}，经络：{经络} ↔ {对应经络}。\n"
-                    f"穴位位置：{穴位详情['位置']}\n"
-                    f"主要功能：{穴位详情['功能']}\n"
-                    f"针刺方法：{穴位详情['针法']}"
-                )
-            }
-        }
-
-    # ========= 灵龟八法 =========
-    @staticmethod
-    def 灵龟八法类(农历, 性别: str):
-        """子午流注类 · 灵龟八法"""
-
-        天干数 = {'甲':9,'己':9,'乙':8,'庚':8,'丙':7,'辛':7,'丁':6,'壬':6,'戊':5,'癸':5}
-        地支数 = {'子':9,'午':9,'丑':8,'未':8,'寅':7,'申':7,'卯':6,'酉':6,'辰':5,'戌':5,'巳':4,'亥':4}
-
-        日干 = 农历.getDayGanExact2()
-        日支 = 农历.getDayZhiExact2()
-        时干 = 农历.getTimeGan()
-        时支 = 农历.getTimeZhi()
-
-        总和 = 天干数[日干] + 地支数[日支] + 天干数[时干] + 地支数[时支]
-
-        初数 = 9 if 日干 in ['甲','丙','戊','庚','壬'] else 6
-        余数 = 总和 % 初数 or 初数
-
-        if 余数 == 5:
-            余数 = 2 if 性别 == "男" else 8
-
-        八法表 = {
-            1:('申脉','阳跷脉','坎','坎一宫'),
-            2:('照海','阴跷脉','坤','坤二宫'),
-            3:('外关','阳维脉','震','震三宫'),
-            4:('足临泣','带脉','巽','巽四宫'),
-            6:('公孙','冲脉','乾','乾六宫'),
-            7:('后溪','督脉','兑','兑七宫'),
-            8:('内关','阴维脉','艮','艮八宫'),
-            9:('列缺','任脉','离','离九宫')
-        }
-
-        穴位, 经络, 卦, 宫位 = 八法表.get(余数)
-        穴位详情 = 子午流注类.穴位详情(穴位)
-        方位, 吉凶 = 子午流注类.八卦信息(卦)
-
-        描述 = 子午流注类.生成描述(农历)
-        摘要 = f"{穴位}（{经络}），属{卦}卦（{方位}），{吉凶}"
-
-        return {
-            "state": 穴位,
-            "attributes": {
-                "数值": 余数,
-                "卦": 卦,
-                "宫位": 宫位,
-                "方位": 方位,
-                "吉凶": 吉凶,
-                "穴位": 穴位,
-                "经络": 经络,
-                "位置": 穴位详情['位置'],
-                "功能": 穴位详情['功能'],
-                "针法": 穴位详情['针法'],
-                "摘要": 摘要,
-                "描述": 描述,
-                "详细描述": (
-                    f"{描述}，灵龟八法得数：{余数}，对应{宫位}（{卦}卦，方位：{方位}，吉凶：{吉凶}）。\n"
-                    f"取穴：{穴位}（{经络}）。\n"
-                    f"穴位位置：{穴位详情['位置']}\n"
-                    f"主要功能：{穴位详情['功能']}\n"
-                    f"针刺方法：{穴位详情['针法']}"
-                )
-            }
-        }
-
-    # ========= 飞腾八法（补配穴） =========
-    @staticmethod
-    def 飞腾八法类(农历):
-        """子午流注类 · 飞腾八法"""
-
-        日干 = 农历.getDayGanExact2()
-        时支 = 农历.getTimeZhi()
-
-        起点 = 子午流注类.飞腾起点.get(日干, "子")
-        时序 = ["子","丑","寅","卯","辰","巳","午","未","申","酉","戌","亥"]
-
-        飞腾数 = (时序.index(时支) - 时序.index(起点)) % 9 + 1
-        卦表 = {1:"坎",2:"坤",3:"震",4:"巽",5:"中",6:"乾",7:"兑",8:"艮",9:"离"}
-        卦 = 卦表[飞腾数]
-        方位, 吉凶 = 子午流注类.八卦信息(卦)
-        # 配穴
-        配穴信息 = 子午流注类.飞腾配穴表.get(飞腾数, (None, None))
-        穴位, 经络 = 配穴信息 if 配穴信息[0] else ("无", "")
-        穴位详情 = 子午流注类.穴位详情(穴位)
-
-        升降表 = {
-            "坎":"宜降","坤":"宜补","震":"宜升","巽":"宜疏散",
-            "乾":"大补","兑":"慎泻","艮":"宜止","离":"宜升散","中":"平补平泻"
-        }
-        升降 = 升降表[卦]
-
-        描述 = 子午流注类.生成描述(农历)
-        摘要 = f"飞腾数：{飞腾数}（{卦}卦，{方位}，{吉凶}），{升降}，配穴：{穴位}({经络})"
-
-        return {
-            "state": 穴位,
-            "attributes": {
-                "飞腾数": 飞腾数,
-                "卦": 卦,
-                "方位": 方位,
-                "吉凶": 吉凶,
-                "升降补泻": 升降,
-                "配穴": 穴位,
-                "配穴经络": 经络,
-                "位置": 穴位详情['位置'],
-                "功能": 穴位详情['功能'],
-                "针法": 穴位详情['针法'],
-                "起点时辰": 起点,
-                "当前时辰": 时支,
-                "摘要": 摘要,
-                "描述": 描述,
-                "详细描述": (
-                    f"{描述}，飞腾八法起点为{起点}时，顺数至{时支}时得数：{飞腾数}。\n"
-                    f"对应卦象：{卦}卦（方位：{方位}，吉凶：{吉凶}）。\n"
-                    f"补泻建议：{升降}。\n"
-                    f"配穴：{穴位}（{经络}），位置：{穴位详情['位置']}，功能：{穴位详情['功能']}，针法：{穴位详情['针法']}"
-                )
-            }
-        }
-
-    # ========= 迎随补泻法 =========
-    @staticmethod
-    def 迎随补泻类(农历, dt):
-        """子午流注类 · 迎随补泻法"""
-
-        小时 = dt.hour
-        时支 = 子午流注类.小时转时支(小时)
-
-        当前经络 = 子午流注类.时辰经络[时支][0]
-        顺经 = 子午流注类.顺经(当前经络)
-        逆经 = 子午流注类.逆经(当前经络)
-
-        补法 = f"顺气补：顺着气血流注方向 → {当前经络} → {顺经}"
-        泻法 = f"逆气泻：逆着气血流注方向 → {当前经络} → {逆经}"
-
-        描述 = 子午流注类.生成描述(农历)
-        摘要 = f"{当前经络}当令，顺气补（{顺经}），逆气泻（{逆经}）"
-
-        return {
-            "state": 当前经络,
-            "attributes": {
-                "时支": 时支,
-                "当前经络": 当前经络,
-                "顺气经络": 顺经,
-                "逆气经络": 逆经,
-                "补法": 补法,
-                "泻法": 泻法,
-                "摘要": 摘要,
-                "描述": 描述,
-                "详细描述": (
-                    f"{描述}，气血流注于{当前经络}。\n"
-                    f"顺气补法：{补法}\n"
-                    f"逆气泻法：{泻法}\n"
-                    f"说明：迎随补泻法以气血流注方向为依据，顺为补，逆为泻。"
-                )
-            }
-        }
-
-# 五运六气类
-class 五运六气类:
-    """五运六气全功能推演类"""
-
-    五运表 = 五运六气映射.get("五运表", {})
-    六气表 = 五运六气映射.get("六气表", {})                     # 年支 -> 司天
-    五运偏盛偏衰 = 五运六气映射.get("五运偏盛偏衰", {})
-    六气偏盛偏衰 = 五运六气映射.get("六气偏盛偏衰", {})
-    六步主气表 = 五运六气映射.get("六步主气", [])               # 步名列表
-    六气轮转表 = 五运六气映射.get("六气轮转", [])               # 三阴三阳固定次序
-    六气属性表 = 五运六气映射.get("六气属性", {})
-    五行关系表 = 五运六气映射.get("五行关系", {})
-    气机病机表 = 五运六气映射.get("气机病机", {})
-    主气六步序列表 = 五运六气映射.get("主气六步序列", [])
-    司天在泉对表 = 五运六气映射.get("司天在泉对", {})
-    年支五行表 = 五运六气映射.get("年支五行", {})
-    三阴三阳顺序表 = 五运六气映射.get("三阴三阳顺序", [])
-    标本中气表 = 五运六气映射.get("标本中气", [])
-
-    # ========= 全局缓存 =========
-    _年度缓存 = {}   # 键："年干+年支" -> 年度数据
-
-    # ========= 基础取数（无缓存，高频调用） =========
-    @staticmethod
-    def 获取五运法(年干: str) -> str:
-        return 五运六气类.五运表.get(年干, "未知")
-
-    @staticmethod
-    def 获取六气法(年支: str) -> str:          # 返回司天
-        return 五运六气类.六气表.get(年支, "未知")
-
-    @staticmethod
-    def 获取在泉法(年支: str) -> str:
-        司天 = 五运六气类.获取六气法(年支)
-        return 五运六气类.司天在泉对表.get(司天, "未知")
-
-    @staticmethod
-    def 获取五运偏差法(五运: str) -> list:
-        return 五运六气类.五运偏盛偏衰.get(五运, ["未知", "未知"])
-
-    @staticmethod
-    def 获取六气偏差法(六气: str) -> list:
-        return 五运六气类.六气偏盛偏衰.get(六气, ["未知", "未知"])
-
-    # ========= 六步主气（固定） =========
-    @staticmethod
-    def 获取六步主气法() -> list:
-        return list(zip(五运六气类.六步主气表, 五运六气类.主气六步序列表))
-
-    # ========= 六步客气（经典司天在泉法） =========
-    @staticmethod
-    def 获取六步客气法(年支: str) -> list:
-        """
-        按年支定司天（三之气）→ 在泉（终之气），
-        依据三阴三阳固定次序推算六步客气。
-        """
-        司天 = 五运六气类.获取六气法(年支)
-        在泉 = 五运六气类.获取在泉法(年支)
-        顺序 = 五运六气类.三阴三阳顺序表
-
-        if 司天 not in 顺序 or 在泉 not in 顺序:
-            return [(步, "未知") for 步 in 五运六气类.六步主气表]
-
-        # 以三之气为司天，逆推初之气、二之气
-        司天位置 = 顺序.index(司天)
-        # 四之气：司天后一位
-        四气位置 = (司天位置 + 1) % 6
-        # 五之气：司天后两位
-        五气位置 = (司天位置 + 2) % 6
-        # 终之气（在泉）应位于司天后三位（对化关系），校验一下
-        终气位置 = (司天位置 + 3) % 6
-        if 顺序[终气位置] != 在泉:
-            # 数据异常时回退为直接使用轮转表
-            终气位置 = 顺序.index(在泉)
-        # 初之气：在泉后一位
-        初气位置 = (终气位置 + 1) % 6
-        # 二之气：在泉后两位
-        二气位置 = (终气位置 + 2) % 6
-
-        # 六步客气序列：初、二、三、四、五、终
-        客气序列 = [
-            顺序[初气位置],
-            顺序[二气位置],
-            顺序[司天位置],
-            顺序[四气位置],
-            顺序[五气位置],
-            顺序[终气位置]
-        ]
-        return list(zip(五运六气类.六步主气表, 客气序列))
-
-    # ========= 气机 -> 病机（基础） =========
-    @staticmethod
-    def 获取病机法(气机: str) -> dict:
-        return 五运六气类.气机病机表.get(气机, {
-            "性质": "中和", "病机": "气机调畅", "症候": "平", "脏腑": "无", "经络": "无"
-        })
-
-    # ========= 运气同化计算 =========
-    @staticmethod
-    def 计算运气同化法(年干: str, 年支: str) -> dict:
-        """返回天符、岁会等特殊年份判定"""
-        岁运 = 五运六气类.获取五运法(年干)       # 如 "木运"
-        岁运五行 = 岁运.replace("运", "")        # "木"
-        司天 = 五运六气类.获取六气法(年支)       # "厥阴风木"
-        在泉 = 五运六气类.获取在泉法(年支)
-        司天五行 = 五运六气类.六气属性表[司天]["五行"]
-        在泉五行 = 五运六气类.六气属性表[在泉]["五行"]
-        年支五行值 = 五运六气类.年支五行表[年支]
-
-        # 天符：岁运五行 == 司天五行
-        天符 = (岁运五行 == 司天五行)
-        # 岁会：岁运五行 == 年支五行
-        岁会 = (岁运五行 == 年支五行值)
-        # 同天符：阳年（甲丙戊庚壬）且岁运太过，岁运五行 == 在泉五行
-        阳干 = set("甲丙戊庚壬")
-        同天符 = (年干 in 阳干) and (岁运五行 == 在泉五行)
-        # 同岁会：阴年（乙丁己辛癸）且岁运不及，岁运五行 == 在泉五行
-        阴干 = set("乙丁己辛癸")
-        同岁会 = (年干 in 阴干) and (岁运五行 == 在泉五行)
-        # 太乙天符：天符 + 岁会
-        太乙天符 = 天符 and 岁会
-
-        描述列表 = []
-        if 天符:
-            描述列表.append("天符")
-        if 岁会:
-            描述列表.append("岁会")
-        if 同天符:
-            描述列表.append("同天符")
-        if 同岁会:
-            描述列表.append("同岁会")
-        if 太乙天符:
-            描述列表.append("太乙天符")
-        描述 = "，".join(描述列表) if 描述列表 else "平气"
-
-        return {
-            "天符": 天符, "岁会": 岁会,
-            "同天符": 同天符, "同岁会": 同岁会,
-            "太乙天符": 太乙天符,
-            "描述": 描述
-        }
-
-    # ========= 胜复推导 =========
-    @staticmethod
-    def 计算胜复法(五运: str, 司天: str) -> dict:
-        """
-        根据岁运太过与司天，给出可能的胜气、复气及影响。
-        简化模型：五运太过则本气胜，所不胜之子来复。
-        """
-        # 五行相胜关系：木胜土，金胜木，火胜金，水胜火，土胜水
-        胜克 = {"木": "土", "火": "金", "土": "水", "金": "木", "水": "火"}
-        # 五行相生关系（子来复）：如木胜土，土之子金复木
-        子复 = {"土": "金", "金": "水", "水": "木", "木": "火", "火": "土"}
-
-        # 判断岁运是否太过
-        太过不及 = 五运六气类.获取五运偏差法(五运)
-        是太过 = "太过" in 太过不及[0]
-
-        if not 是太过:
-            return {"胜气": "无（岁运不及）", "复气": "无", "病机提示": "胜复不显"}
-
-        岁运五行 = 五运.replace("运", "")
-        所胜 = 胜克.get(岁运五行)
-        复气五行 = 子复.get(所胜) if 所胜 else None
-
-        司天气机 = 五运六气类.六气属性表[司天]["五行"] if 司天 in 五运六气类.六气属性表 else None
-
-        # 如果司天五行与岁运五行同，则胜气更烈
-        加临 = "司天同化，胜气倍增" if 司天气机 == 岁运五行 else "司天不同化"
-
-        复气描述 = f"{复气五行}气来复" if 复气五行 else "复气不明"
-        return {
-            "胜气": f"{岁运五行}气偏胜",
-            "所胜": f"{所胜}气受抑",
-            "复气": f"{复气五行}气来复（{所胜}之子）" if 复气五行 else "待考",
-            "病机提示": f"{岁运五行}胜乘{所胜}，{加临}，{复气描述}。",
-        }
-
-    # ========= 客主相交（升级版：含标本中气） =========
-    @staticmethod
-    def 客主相交法(客气: str, 主气: str) -> dict:
-        # 同气相求
-        if 客气 == 主气:
-            气机 = 五运六气类.六气属性表[客气]["气机"]
-            病机 = 五运六气类.获取病机法(气机)
-            标本 = 五运六气类.标本中气表[客气]
-        
-            return {
-                "关系": "同气相求",
-                "五行关系": "同类相应",
-                "病机": 病机["病机"],
-                "症候": 病机["症候"],
-                "脏腑": 病机["脏腑"],
-                "经络": 病机["经络"],
-                "标本从化": 标本["从化"],
-                "摘要": f"{客气} 同气相求 → {病机['病机']}"
-            }
-
-        客五行 = 五运六气类.六气属性表[客气]["五行"]
-        主五行 = 五运六气类.六气属性表[主气]["五行"]
-        客标本 = 五运六气类.标本中气表[客气]
-        主标本 = 五运六气类.标本中气表[主气]
-
-        # 君臣顺逆：主为君，客为臣。
-        生克关系 = None
-        原始方向 = None
-        正向键 = f"{客五行},{主五行}"
-        反向键 = f"{主五行},{客五行}"
-        if 正向键 in 五运六气类.五行关系表:
-            生克关系 = 五运六气类.五行关系表[正向键]
-            原始方向 = "客→主"
-        elif 反向键 in 五运六气类.五行关系表:
-            生克关系 = 五运六气类.五行关系表[反向键]
-            原始方向 = "主→客"
-            if "生" in 生克关系:
-                生克关系 = 生克关系.replace("生", "被生")
-            elif "克" in 生克关系:
-                生克关系 = 生克关系.replace("克", "被克")
-
-        # 相得/不相得判定
-        if not 生克关系:
-            顺逆 = "未知"
-        elif "生" in 生克关系:
-            if 原始方向 == "客→主":
-                顺逆 = "相得（客气生主气）"
-            else:
-                顺逆 = "不相得（主气生客气）"
-        else:  # 克
-            if 原始方向 == "客→主":
-                顺逆 = "不相得（客气克主气，逆）"
-            else:
-                顺逆 = "相得（主气克客气，君胜臣）"
-
-        # 偏盛乘侮
-        客太过, _ = 五运六气类.获取六气偏差法(客气)
-        主太过, _ = 五运六气类.获取六气偏差法(主气)
-        客偏盛 = "太过" in 客太过
-        主偏盛 = "太过" in 主太过
-
-        乘侮 = ""
-        if 客偏盛 and "克" in (生克关系 or ""):
-            if "被克" in 生克关系:   # 主克客，但客偏盛反侮
-                乘侮 = "侮"
-            else:                     # 客克主，乘
-                乘侮 = "乘"
-        elif 主偏盛 and "克" in (生克关系 or ""):
-            if "被克" in 生克关系:   # 客克主，主偏盛反侮
-                乘侮 = "侮"
-            else:                     # 主克客，乘
-                乘侮 = "乘"
-
-        # 主导气机
-        if 客偏盛 and not 主偏盛:
-            主导气机 = 五运六气类.六气属性表[客气]["气机"]
-            主导类型 = "客气偏盛"
-        elif 主偏盛 and not 客偏盛:
-            主导气机 = 五运六气类.六气属性表[主气]["气机"]
-            主导类型 = "主气偏盛"
-        elif 客偏盛 and 主偏盛:
-            # 两盛则看生克：客胜主为从，较顺；主胜客为逆
-            if "被克" in (生克关系 or ""):
-                主导气机 = 五运六气类.六气属性表[客气]["气机"]
-                主导类型 = "客主俱盛，客气反胜（从）"
-            else:
-                主导气机 = 五运六气类.六气属性表[主气]["气机"]
-                主导类型 = "客主俱盛，主气胜客（逆）"
-        else:
-            # 皆不偏盛，按常规胜气
-            if "被克" in (生克关系 or ""):
-                主导气机 = 五运六气类.六气属性表[客气]["气机"]
-                主导类型 = "客气胜主（从）"
-            else:
-                主导气机 = 五运六气类.六气属性表[主气]["气机"]
-                主导类型 = "主气胜客（逆）"
-
-        病机 = 五运六气类.获取病机法(主导气机)
-        五行描述 = f"{生克关系 or '平气'}{'（' + 乘侮 + '）' if 乘侮 else ''}"
-
-        # 标本从化提示
-        标本提示 = f"客从{客标本['从化']}，主从{主标本['从化']}"
-
-        return {
-            "关系": "客主相交",
-            "五行关系": 五行描述,
-            "顺逆": 顺逆,
-            "主导类型": 主导类型,
-            "主导气机": 主导气机,
-            "病机": 病机["病机"],
-            "症候": 病机["症候"],
-            "脏腑": 病机["脏腑"],
-            "经络": 病机["经络"],
-            "标本从化": 标本提示,
-            "摘要": f"{客气} vs {主气}：{五行描述} {顺逆} {主导类型} → {病机['病机']}"
-        }
-
-    # ========= 六步运气数据 =========
-    @staticmethod
-    def 获取当前步序法(农历) -> int:
-        """判定当前处于六气中的第几步 (0-5)"""
-        节点 = ["大寒", "春分", "小满", "大暑", "秋分", "小雪"]
-        当前阳历 = 农历.getSolar().toYmd()
-        节气表 = 农历.getJieQiTable()
-        # 从最后一步往回查
-        for 序号 in range(5, -1, -1):
-            目标 = 节气表.get(节点[序号])
-            if 目标 and 当前阳历 >= 目标.toYmd():
-                return 序号
-        return 5  # 兜底
-
-    @staticmethod
-    def 六步运气数据类法(年干: str, 年支: str) -> list:
-        """
-        全量排盘计算：生成全年 6 个步序的详细数据列表
-        """
-        主气序列 = 五运六气类.获取六步主气法()
-        客气序列 = 五运六气类.获取六步客气法(年支) # 遵循要求：改用年支计算客气起点
-
-        结果 = []
-        for (步, 主气), (_, 客气) in zip(主气序列, 客气序列):
-            关系 = 五运六气类.客主相交法(客气, 主气)
-            结果.append({
-                "state": 步,
-                "attributes": {
-                    "主气": 主气,
-                    "客气": 客气,
-                    "五行关系": 关系["五行关系"],
-                    "顺逆": 关系.get("顺逆", ""),
-                    "主导类型": 关系.get("主导类型", ""),
-                    "病机": 关系["病机"],
-                    "症候": 关系["症候"],
-                    "脏腑": 关系["脏腑"],
-                    "经络": 关系["经络"],
-                    "标本从化": 关系["标本从化"],
-                    # 跨类联动：获取对应的方剂参考
-                    "辅行诀参考": 辅行诀脏腑用药法要类.五运六气病机辅行诀选方(关系["病机"]),
-                    "经方参考": 伤寒杂病论类.五运六气病机伤寒选方(关系["病机"]),
-                    "摘要": f"{步}：{关系['摘要']}"
-                }
-            })
-        return 结果
-
-    # ========= 年度五运 =========
-    @staticmethod
-    def 年五运类法(年干: str) -> dict:
-        五运 = 五运六气类.获取五运法(年干)
-        太过, 不及 = 五运六气类.获取五运偏差法(五运)
-        摘要 = f"{五运}（{太过} / {不及}）"
-        
-        return {
-            "state": f"{五运}{'太过' if '太过' in 太过 else '不及'}",
-            "attributes": {
-                "五运": 五运, 
-                "太过": 太过, 
-                "不及": 不及,
-                "摘要": 摘要
-            }
-        }
-
-    # ========= 年度六气 =========
-    @staticmethod
-    def 年六气类法(年支: str) -> dict:
-        司天 = 五运六气类.获取六气法(年支)
-        在泉 = 五运六气类.获取在泉法(年支)
-        太过, 不及 = 五运六气类.获取六气偏差法(司天)
-        摘要 = f"{司天}（{太过} / {不及}）"
-
-        return {
-            "state": f"{司天}司天·{在泉}在泉",
-            "attributes": {
-                "司天": 司天, 
-                "在泉": 在泉, 
-                "司天太过": 太过, 
-                "司天不及": 不及,
-                "摘要": 摘要
-                }
-        }
-
-    # ========= 年度总览（含运气同化、胜复） =========
-    @staticmethod
-    def 年度总览类(年干: str, 年支: str) -> dict:
-        五运 = 五运六气类.年五运类法(年干)
-        六气 = 五运六气类.年六气类法(年支)
-        六步 = 五运六气类.六步运气数据类法(年干, 年支)
-        同化 = 五运六气类.计算运气同化法(年干, 年支)
-        胜复 = 五运六气类.计算胜复法(五运["attributes"]["五运"], 六气["attributes"]["司天"])
-
-        扁平属性 = {}
-        # 平铺五运
-        扁平属性["岁运"] = 五运["attributes"]["五运"]
-        扁平属性["岁运太过"] = 五运["attributes"]["太过"]
-        扁平属性["岁运不及"] = 五运["attributes"]["不及"]
-        扁平属性["岁运摘要"] = 五运["attributes"]["摘要"]
-        # 平铺六气
-        扁平属性["司天"] = 六气["attributes"]["司天"]
-        扁平属性["在泉"] = 六气["attributes"]["在泉"]
-        扁平属性["司天太过"] = 六气["attributes"]["司天太过"]
-        扁平属性["司天不及"] = 六气["attributes"]["司天不及"]
-        扁平属性["司天摘要"] = 六气["attributes"]["摘要"]
-        # 平铺同化
-        扁平属性["运气同化类型"] = 同化["描述"]
-        扁平属性["是否为天符"] = 同化["天符"]
-        扁平属性["是否为岁会"] = 同化["岁会"]
-        扁平属性["是否为同天符"] = 同化["同天符"]
-        扁平属性["是否为同岁会"] = 同化["同岁会"]
-        扁平属性["是否为太乙天符"] = 同化["太乙天符"]
-        # 平铺胜复
-        扁平属性["胜气"] = 胜复.get("胜气", "")
-        扁平属性["所胜受抑"] = 胜复.get("所胜", "")
-        扁平属性["复气"] = 胜复.get("复气", "")
-        扁平属性["胜复病机提示"] = 胜复.get("病机提示", "")
-        # 平铺六步摘要（每步一行）
-        六步摘要列表 = []
-        for 步数据 in 六步:
-            步名 = 步数据["state"]
-            主气 = 步数据["attributes"]["主气"]
-            客气 = 步数据["attributes"]["客气"]
-            病机 = 步数据["attributes"]["病机"]
-            六步摘要列表.append(f"{步名}：主气{主气}，客气{客气}，病机{病机}")
-        扁平属性["六步摘要"] = "；".join(六步摘要列表)
-        # 添加年度摘要
-        扁平属性["年度摘要"] = f"{年干}{年支}年：{五运['state']}，{六气['state']}。{同化['描述']}。{胜复['病机提示']}"
-
-        return {
-            "state": f"{年干}{年支}年运气总览",
-            "attributes": 扁平属性
-        }
-
-    # ========= 全量计算（带缓存） =========
-    @staticmethod
-    def 全量计算类(农历) -> dict:
-        年干 = 农历.getYearGanExact()
-        年支 = 农历.getYearZhiExact()
-        缓存键 = f"{年干}{年支}"
-
-        if 缓存键 in 五运六气类._年度缓存:
-            年度数据 = 五运六气类._年度缓存[缓存键]
-        else:
-            年度数据 = {
-                "年五运数据": 五运六气类.年五运类法(年干),
-                "年六气数据": 五运六气类.年六气类法(年支),
-                "六步运气数据_全量": 五运六气类.六步运气数据类法(年干, 年支),
-                "年度总览数据": 五运六气类.年度总览类(年干, 年支)
-            }
-            五运六气类._年度缓存[缓存键] = 年度数据
-
-        # 取出当前步
-        全量六步 = 五运六气类.六步运气数据类法(年干, 年支)
-        当前索引 = 五运六气类.获取当前步序法(农历)
-        当前步 = 全量六步[当前索引]
-
-        return {
-            "年五运数据": 年度数据["年五运数据"],
-            "年六气数据": 年度数据["年六气数据"],
-            "六步运气数据": {
-                "state": 当前步["state"],
-                "attributes": 当前步["attributes"]
-            },
-            "年度总览数据": 年度数据["年度总览数据"]
-        }
-
-# 辅行诀脏腑用药法要类
-class 辅行诀脏腑用药法要类:
-    """辅行诀脏腑用药法要 · 症状选方（直接输出方剂）"""
-
-    汤液表 = 岐黄类映射.get("辅行诀汤液表", {})
-    辨证规则表 = 岐黄类映射.get("辅行诀辨证规则表", {})
-
-    # ========= 供前端选择框 =========
-    @staticmethod
-    def 获取所有大类法():
-        """第一级下拉框：包含五脏、虚劳、救逆"""
-        return ["肝", "心", "脾", "肺", "肾", "劳", "误治"]
-
-    @staticmethod
-    def 获取大类症状法(选中大类: str):
-        """
-        第二级下拉框：根据大类动态过滤症状
-        """
-        症状集 = set()
-        for 证型, 规则 in 辅行诀脏腑用药法要类.辨证规则表.items():
-            # 逻辑：
-            # 1. 如果选的是“劳”，匹配所有带“劳”字的证型 (如：肝劳、心劳)
-            # 2. 如果选的是“误治”，匹配所有带“误治”字的证型 (如：误治伤肝)
-            # 3. 如果选的是脏腑，匹配该脏腑下实、虚的证型 (如：肝实、肝虚)
-            if 选中大类 == "劳" and "劳" in 证型:
-                症状集.update(规则.get("症状", []))
-            elif 选中大类 == "误治" and "误治" in 证型:
-                症状集.update(规则.get("症状", []))
-            elif 选中大类 in 证型 and "劳" not in 证型 and "误治" not in 证型:
-                # 排除掉劳和误治，只留纯脏腑虚实
-                症状集.update(规则.get("症状", []))
-        
-        # 修正：必须返回处理后的列表
-        return sorted(list(症状集))
-    
-    # 统一选方入口
-    @staticmethod
-    def 辅行诀选方类(症状):
-        """症状 (str 或 list) → 直接输出方剂实体"""
-        if not 症状:
-            return {"state": "未选症状", "summary": "请选择症状", "attributes": {}}
-
-        if isinstance(症状, str):
-            症状列表 = [症状]
-        else:
-            症状列表 = 症状
-
-        # 匹配证型
-        最佳证型 = None
-        最高命中 = 0
-        匹配方名 = None
-
-        for 证型, 规则 in 辅行诀脏腑用药法要类.辨证规则表.items():
-            命中 = sum(1 for s in 症状列表 if s in 规则["症状"])
-            if 命中 > 最高命中:
-                最高命中 = 命中
-                最佳证型 = 证型
-                匹配方名 = 规则.get("方剂") # 修正：直接从规则里拿正确方名
-
-        if not 最佳证型:
-            return {"state": "未匹配", "summary": "未找到对应方剂", "attributes": {}}
-
-        # 判定脏腑（用于去汤液表查找详情）
-        # 辅行诀汤液表的一级 Key 是 ["肝", "心", "脾", "肺", "肾"]
-        脏腑 = "肝" 
-        for 关键字 in ["肝", "心", "脾", "肺", "肾"]:
-            if 关键字 in 最佳证型:
-                脏腑 = 关键字
-                break
-
-        # 判定虚实性质及“查询分类”
-        # 我们定义一个“查询分类”变量，专门用来对应 汤液表 的一级 Key
-        查询分类 = 脏腑
-        虚实性质 = "实" if "实" in 最佳证型 else "虚"
-
-        if "劳" in 最佳证型:
-            查询分类 = "救诸劳损"  # 对应 JSON 里的一级 Key
-            虚实性质 = "劳损"
-        elif "误治" in 最佳证型:
-            查询分类 = "救误治"    # 对应 JSON 里的一级 Key
-            虚实性质 = "救逆"
-
-        # 取方剂直接输出
-        # 注意：这里传进去的第一个参数必须是“查询分类”
-        return 辅行诀脏腑用药法要类._辅行诀输出方剂(查询分类, 虚实性质, 最佳证型, 匹配方名)
-
-    # ========= 输出方剂（统一出口） =========
-    @staticmethod
-    def _辅行诀输出方剂(脏, 虚实, 证型, 方名):
-        # 增加防御逻辑，防止 Key 错误
-        try:
-            方剂 = 辅行诀脏腑用药法要类.汤液表[脏][方名]
-        except KeyError:
-            return {"state": "数据缺失", "summary": f"库中未找到 {方名}", "attributes": {}}
-
-        组成 = "、".join(方剂["组成"])
-        # 加减条目处理
-        加减条目 = 方剂.get("加减", [])
-        加减文本 = "；".join(f"{条['证']}：加{', '.join(条['加'])}" for 条 in 加减条目) if 加减条目 else "无特定加减"
-
-        return {
-            "state": f"{方名}",
-            "summary": f"{方名} ({脏}{虚实})：{方剂.get('主治')}",
-            "attributes": {
-                "证型": 证型,
-                "方名": 方名,
-                "组成": 组成,
-                "主治": 方剂["主治"],
-                "加减参考": 加减文本,
-                "脏腑": 脏,
-                "虚实性质": 虚实
-            }
-        }
-
-    # 子午流注联动
-    @staticmethod
-    def 子午流注纳子法辅行诀选方(时支经络, 证型="平"):
-        经络到脏 = {
-            "肺经":"肺","大肠经":"肺","胃经":"脾","脾经":"脾",
-            "心经":"心","小肠经":"心","膀胱经":"肾","肾经":"肾",
-            "心包经":"心","三焦经":"心","胆经":"肝","肝经":"肝"
-        }
-        经络 = 时支经络.split()[-1]
-        脏 = 经络到脏[经络]
-        虚实 = "实" if 证型 == "实" else "虚"
-        方名 = f"小泻{脏}汤" if 虚实 == "实" else f"小补{脏}汤"
-        方 = 辅行诀脏腑用药法要类.汤液表[脏][方名]
-        return f"{方名}：{'、'.join(方['组成'])}"
-
-    # 五运六气联动
-    @staticmethod
-    def 五运六气病机辅行诀选方(病机):
-        映射 = [
-            ("肝",["风动","不舒","亢盛","火旺"],"实"),
-            ("肝",["不足","虚寒","血亏"],"虚"),
-            ("心",["火旺","偏亢","不宁","上炎"],"实"),
-            ("心",["不足","虚烦","阳虚"],"虚"),
-            ("脾",["湿困","壅滞","实满"],"实"),
-            ("脾",["虚弱","不运","虚寒"],"虚"),
-            ("肺",["燥伤","气逆","痰盛"],"实"),
-            ("肺",["气虚","不足","虚劳"],"虚"),
-            ("肾",["寒凝","水停","热结"],"实"),
-            ("肾",["不足","虚衰","亏虚"],"虚"),
-        ]
-        for 脏, 关键词, 虚实 in 映射:
-            if 脏 in 病机 and any(k in 病机 for k in 关键词):
-                方名 = f"小泻{脏}汤" if 虚实=="实" else f"小补{脏}汤"
-                方 = 辅行诀脏腑用药法要类.汤液表[脏][方名]
-                return f"{方名}：{'、'.join(方['组成'])}"
-        return ""
-
-
-class 伤寒杂病论类:
-    """伤寒杂病论 · 动态索引与自动辨证引擎"""
-
-    方剂库 = 岐黄类映射.get("伤寒杂病论方剂库", {})
-    六经列表 = ["太阳", "阳明", "少阳", "太阴", "少阴", "厥阴"]
-
-    # 动态索引（供 UI 选择器使用）
-    @staticmethod
-    def 获取所有六经法():
-        """直接返回硬编码的列表"""
-        return 伤寒杂病论类.六经列表
-
-    @staticmethod
-    def 获取经下所有证型法(选中六经: str):
-        """提取该经下所有的证型标识"""
-        return sorted(list(set(
-            v.get("证型") for v in 伤寒杂病论类.方剂库.values() 
-            if v.get("六经") == 选中六经 and v.get("证型")
-        )))
-
-    @staticmethod
-    def 获取证型下所有方名法(选中证型: str):
-        """提取对应证型的方名（通常为1个，支持多个）"""
-        return sorted([
-            k for k, v in 伤寒杂病论类.方剂库.items() 
-            if v.get("证型") == 选中证型
-        ])
-
-    @staticmethod
-    def 根据证型查方名法(目标证型: str) -> str | None:
-        """【核心优化】不再查规则表，直接在库中反向搜索证型对应的方名"""
-        for 方名, 详情 in 伤寒杂病论类.方剂库.items():
-            if 详情.get("证型") == 目标证型:
-                return 方名
-        return None
-
-    @staticmethod
-    def 获取方剂数据类(方名: str):
-        """统一输出接口：输入方名，输出符合 Sensor 标准的字典"""
-        if not 方名:
-            return {"state": "未选择", "summary": "待查", "attributes": {}}
-            
-        详情 = 伤寒杂病论类.方剂库.get(方名)
-        if not 详情:
-            return {"state": "缺失", "summary": f"库中无【{方名}】", "attributes": {}}
-
-        return {
-            "state": 方名,
-            "summary": f"{方名}：{详情.get('功用', '详见属性')}",
-            "attributes": {
-                "组成": "、".join(详情.get("组成", [])),
-                "主治": 详情.get("主治"),
-                "功用": 详情.get("功用"),
-                "六经": 详情.get("六经"),
-                "辨证": 详情.get("证型")
-            }
-        }
-
-    # 子午流注纳子法伤寒选方
-    @staticmethod
-    def 子午流注纳子法伤寒选方(时支经络: str, 证型: str = "平") -> str:
-        """当令经络自动推方"""
-        经络到六经 = {
-            "肺经":"太阴","大肠经":"阳明","胃经":"阳明","脾经":"太阴",
-            "心经":"少阴","小肠经":"太阳","膀胱经":"太阳","肾经":"少阴",
-            "心包经":"厥阴","三焦经":"少阳","胆经":"少阳","肝经":"厥阴"
-        }
-        经络名 = 时支经络.split(" ")[-1] if " " in 时支经络 else 时支经络
-        六经 = 经络到六经.get(经络名)
-        if not 六经: return ""
-
-        # 精确映射到库里的证型字符串
-        映射 = {
-            "太阳": {"实": "太阳-表寒实", "虚": "太阳-表虚寒", "平": "太阳-表虚寒"},
-            "阳明": {"实": "阳明-里热实重", "虚": "阳明-里热盛伤津", "平": "阳明-里热盛"},
-            "少阳": {"实": "少阳-半表半里兼阳明", "虚": "少阳-半表半里", "平": "少阳-半表半里"},
-            "太阴": {"实": "太阴-里虚寒", "虚": "太阴-里虚寒", "平": "太阴-里虚寒"},
-            "少阴": {"实": "少阴-里虚寒", "虚": "少阴-里虚寒", "平": "少阴-里虚寒"},
-            "厥阴": {"实": "厥阴-上热下寒", "虚": "厥阴-血虚寒厥", "平": "厥阴-上热下寒"}
-        }
-        
-        目标证型 = 映射[六经][证型]
-        匹配方名 = 伤寒杂病论类.根据证型查方名法(目标证型)
-        
-        if 匹配方名:
-            方 = 伤寒杂病论类.方剂库[匹配方名]
-            return f"【{匹配方名}】{'、'.join(方['组成'])}"
-        return ""
-    
-    # 五运六气病机伤寒选方
-    @staticmethod
-    def 五运六气病机伤寒选方(病机: str) -> str:
-        """五运六气病机自动推方"""
-        配置表 = [
-            ("肝", ["风动","不舒","亢盛","火旺"], "厥阴-上热下寒"),
-            ("肝", ["不足","虚寒","血亏"], "厥阴-血虚寒厥"),
-            ("心", ["火旺","偏亢","不宁","上炎"], "少阴-里热实"), # 假设库里有此证型
-            ("心", ["不足","虚烦","阳虚"], "少阴-里虚寒"),
-            ("脾", ["湿困","壅滞","实满"], "太阴-里虚寒"), # 太阴以温中为主
-            ("肺", ["燥伤","气逆","痰盛"], "阳明-里热盛"),
-            ("肾", ["寒凝","水停","热结"], "少阴-里虚寒"),
-        ]
-        for 脏, 关键词, 对应证型 in 配置表:
-            if 脏 in 病机 and any(k in 病机 for k in 关键词):
-                匹配方名 = 伤寒杂病论类.根据证型查方名法(对应证型)
-                if 匹配方名:
-                    方 = 伤寒杂病论类.方剂库[匹配方名]
-                    return f"【{匹配方名}】{'、'.join(方['组成'])}"
-        return ""
