+++ custom_components/tianyuan_calendar/tianyuan/qihuang/__init__.py
@@ -1,0 +1,12 @@
+"""天元岐黄引擎."""
+from .ziwuliu import 子午流注类
+from .wuyun_liuqi import 五运六气类
+from .fuxingjue import 辅行诀脏腑用药法要类
+from .shanghan import 伤寒杂病论类
+
+__all__ = [
+    "子午流注类",
+    "五运六气类",
+    "辅行诀脏腑用药法要类",
+    "伤寒杂病论类",
+]
