+++ custom_components/tianyuan_calendar/tianyuan/qihuang/fuxingjue.py
@@ -1,0 +1,157 @@
+"""天元岐黄引擎 - 辅行诀脏腑用药法要."""
+from __future__ import annotations
+
+from ..library import 加载全量映射表类
+
+maps = 加载全量映射表类()
+岐黄类映射 = maps.get("岐黄类", {})
+
+class 辅行诀脏腑用药法要类:
+    """辅行诀脏腑用药法要 · 症状选方（直接输出方剂）"""
+
+    汤液表 = 岐黄类映射.get("辅行诀汤液表", {})
+    辨证规则表 = 岐黄类映射.get("辅行诀辨证规则表", {})
+
+    # ========= 供前端选择框 =========
+    @staticmethod
+    def 获取所有大类法():
+        """第一级下拉框：包含五脏、虚劳、救逆"""
+        return ["肝", "心", "脾", "肺", "肾", "劳", "误治"]
+
+    @staticmethod
+    def 获取大类症状法(选中大类: str):
+        """
+        第二级下拉框：根据大类动态过滤症状
+        """
+        症状集 = set()
+        for 证型, 规则 in 辅行诀脏腑用药法要类.辨证规则表.items():
+            # 逻辑：
+            # 1. 如果选的是“劳”，匹配所有带“劳”字的证型 (如：肝劳、心劳)
+            # 2. 如果选的是“误治”，匹配所有带“误治”字的证型 (如：误治伤肝)
+            # 3. 如果选的是脏腑，匹配该脏腑下实、虚的证型 (如：肝实、肝虚)
+            if 选中大类 == "劳" and "劳" in 证型:
+                症状集.update(规则.get("症状", []))
+            elif 选中大类 == "误治" and "误治" in 证型:
+                症状集.update(规则.get("症状", []))
+            elif 选中大类 in 证型 and "劳" not in 证型 and "误治" not in 证型:
+                # 排除掉劳和误治，只留纯脏腑虚实
+                症状集.update(规则.get("症状", []))
+        
+        # 修正：必须返回处理后的列表
+        return sorted(list(症状集))
+    
+    # 统一选方入口
+    @staticmethod
+    def 辅行诀选方类(症状):
+        """症状 (str 或 list) → 直接输出方剂实体"""
+        if not 症状:
+            return {"state": "未选症状", "summary": "请选择症状", "attributes": {}}
+
+        if isinstance(症状, str):
+            症状列表 = [症状]
+        else:
+            症状列表 = 症状
+
+        # 匹配证型
+        最佳证型 = None
+        最高命中 = 0
+        匹配方名 = None
+
+        for 证型, 规则 in 辅行诀脏腑用药法要类.辨证规则表.items():
+            命中 = sum(1 for s in 症状列表 if s in 规则["症状"])
+            if 命中 > 最高命中:
+                最高命中 = 命中
+                最佳证型 = 证型
+                匹配方名 = 规则.get("方剂") # 修正：直接从规则里拿正确方名
+
+        if not 最佳证型:
+            return {"state": "未匹配", "summary": "未找到对应方剂", "attributes": {}}
+
+        # 判定脏腑（用于去汤液表查找详情）
+        # 辅行诀汤液表的一级 Key 是 ["肝", "心", "脾", "肺", "肾"]
+        脏腑 = "肝" 
+        for 关键字 in ["肝", "心", "脾", "肺", "肾"]:
+            if 关键字 in 最佳证型:
+                脏腑 = 关键字
+                break
+
+        # 判定虚实性质及“查询分类”
+        # 我们定义一个“查询分类”变量，专门用来对应 汤液表 的一级 Key
+        查询分类 = 脏腑
+        虚实性质 = "实" if "实" in 最佳证型 else "虚"
+
+        if "劳" in 最佳证型:
+            查询分类 = "救诸劳损"  # 对应 JSON 里的一级 Key
+            虚实性质 = "劳损"
+        elif "误治" in 最佳证型:
+            查询分类 = "救误治"    # 对应 JSON 里的一级 Key
+            虚实性质 = "救逆"
+
+        # 取方剂直接输出
+        # 注意：这里传进去的第一个参数必须是“查询分类”
+        return 辅行诀脏腑用药法要类._辅行诀输出方剂(查询分类, 虚实性质, 最佳证型, 匹配方名)
+
+    # ========= 输出方剂（统一出口） =========
+    @staticmethod
+    def _辅行诀输出方剂(脏, 虚实, 证型, 方名):
+        # 增加防御逻辑，防止 Key 错误
+        try:
+            方剂 = 辅行诀脏腑用药法要类.汤液表[脏][方名]
+        except KeyError:
+            return {"state": "数据缺失", "summary": f"库中未找到 {方名}", "attributes": {}}
+
+        组成 = "、".join(方剂["组成"])
+        # 加减条目处理
+        加减条目 = 方剂.get("加减", [])
+        加减文本 = "；".join(f"{条['证']}：加{', '.join(条['加'])}" for 条 in 加减条目) if 加减条目 else "无特定加减"
+
+        return {
+            "state": f"{方名}",
+            "summary": f"{方名} ({脏}{虚实})：{方剂.get('主治')}",
+            "attributes": {
+                "证型": 证型,
+                "方名": 方名,
+                "组成": 组成,
+                "主治": 方剂["主治"],
+                "加减参考": 加减文本,
+                "脏腑": 脏,
+                "虚实性质": 虚实
+            }
+        }
+
+    # 子午流注联动
+    @staticmethod
+    def 子午流注纳子法辅行诀选方(时支经络, 证型="平"):
+        经络到脏 = {
+            "肺经":"肺","大肠经":"肺","胃经":"脾","脾经":"脾",
+            "心经":"心","小肠经":"心","膀胱经":"肾","肾经":"肾",
+            "心包经":"心","三焦经":"心","胆经":"肝","肝经":"肝"
+        }
+        经络 = 时支经络.split()[-1]
+        脏 = 经络到脏[经络]
+        虚实 = "实" if 证型 == "实" else "虚"
+        方名 = f"小泻{脏}汤" if 虚实 == "实" else f"小补{脏}汤"
+        方 = 辅行诀脏腑用药法要类.汤液表[脏][方名]
+        return f"{方名}：{'、'.join(方['组成'])}"
+
+    # 五运六气联动
+    @staticmethod
+    def 五运六气病机辅行诀选方(病机):
+        映射 = [
+            ("肝",["风动","不舒","亢盛","火旺"],"实"),
+            ("肝",["不足","虚寒","血亏"],"虚"),
+            ("心",["火旺","偏亢","不宁","上炎"],"实"),
+            ("心",["不足","虚烦","阳虚"],"虚"),
+            ("脾",["湿困","壅滞","实满"],"实"),
+            ("脾",["虚弱","不运","虚寒"],"虚"),
+            ("肺",["燥伤","气逆","痰盛"],"实"),
+            ("肺",["气虚","不足","虚劳"],"虚"),
+            ("肾",["寒凝","水停","热结"],"实"),
+            ("肾",["不足","虚衰","亏虚"],"虚"),
+        ]
+        for 脏, 关键词, 虚实 in 映射:
+            if 脏 in 病机 and any(k in 病机 for k in 关键词):
+                方名 = f"小泻{脏}汤" if 虚实=="实" else f"小补{脏}汤"
+                方 = 辅行诀脏腑用药法要类.汤液表[脏][方名]
+                return f"{方名}：{'、'.join(方['组成'])}"
+        return ""
