+++ custom_components/tianyuan_calendar/tianyuan/qihuang/shanghan.py
@@ -1,0 +1,116 @@
+"""天元岐黄引擎 - 伤寒杂病论."""
+from __future__ import annotations
+
+from ..library import 加载全量映射表类
+
+maps = 加载全量映射表类()
+岐黄类映射 = maps.get("岐黄类", {})
+
+class 伤寒杂病论类:
+    """伤寒杂病论 · 动态索引与自动辨证引擎"""
+
+    方剂库 = 岐黄类映射.get("伤寒杂病论方剂库", {})
+    六经列表 = ["太阳", "阳明", "少阳", "太阴", "少阴", "厥阴"]
+
+    # 动态索引（供 UI 选择器使用）
+    @staticmethod
+    def 获取所有六经法():
+        """直接返回硬编码的列表"""
+        return 伤寒杂病论类.六经列表
+
+    @staticmethod
+    def 获取经下所有证型法(选中六经: str):
+        """提取该经下所有的证型标识"""
+        return sorted(list(set(
+            v.get("证型") for v in 伤寒杂病论类.方剂库.values() 
+            if v.get("六经") == 选中六经 and v.get("证型")
+        )))
+
+    @staticmethod
+    def 获取证型下所有方名法(选中证型: str):
+        """提取对应证型的方名（通常为1个，支持多个）"""
+        return sorted([
+            k for k, v in 伤寒杂病论类.方剂库.items() 
+            if v.get("证型") == 选中证型
+        ])
+
+    @staticmethod
+    def 根据证型查方名法(目标证型: str) -> str | None:
+        """【核心优化】不再查规则表，直接在库中反向搜索证型对应的方名"""
+        for 方名, 详情 in 伤寒杂病论类.方剂库.items():
+            if 详情.get("证型") == 目标证型:
+                return 方名
+        return None
+
+    @staticmethod
+    def 获取方剂数据类(方名: str):
+        """统一输出接口：输入方名，输出符合 Sensor 标准的字典"""
+        if not 方名:
+            return {"state": "未选择", "summary": "待查", "attributes": {}}
+            
+        详情 = 伤寒杂病论类.方剂库.get(方名)
+        if not 详情:
+            return {"state": "缺失", "summary": f"库中无【{方名}】", "attributes": {}}
+
+        return {
+            "state": 方名,
+            "summary": f"{方名}：{详情.get('功用', '详见属性')}",
+            "attributes": {
+                "组成": "、".join(详情.get("组成", [])),
+                "主治": 详情.get("主治"),
+                "功用": 详情.get("功用"),
+                "六经": 详情.get("六经"),
+                "辨证": 详情.get("证型")
+            }
+        }
+
+    # 子午流注纳子法伤寒选方
+    @staticmethod
+    def 子午流注纳子法伤寒选方(时支经络: str, 证型: str = "平") -> str:
+        """当令经络自动推方"""
+        经络到六经 = {
+            "肺经":"太阴","大肠经":"阳明","胃经":"阳明","脾经":"太阴",
+            "心经":"少阴","小肠经":"太阳","膀胱经":"太阳","肾经":"少阴",
+            "心包经":"厥阴","三焦经":"少阳","胆经":"少阳","肝经":"厥阴"
+        }
+        经络名 = 时支经络.split(" ")[-1] if " " in 时支经络 else 时支经络
+        六经 = 经络到六经.get(经络名)
+        if not 六经: return ""
+
+        # 精确映射到库里的证型字符串
+        映射 = {
+            "太阳": {"实": "太阳-表寒实", "虚": "太阳-表虚寒", "平": "太阳-表虚寒"},
+            "阳明": {"实": "阳明-里热实重", "虚": "阳明-里热盛伤津", "平": "阳明-里热盛"},
+            "少阳": {"实": "少阳-半表半里兼阳明", "虚": "少阳-半表半里", "平": "少阳-半表半里"},
+            "太阴": {"实": "太阴-里虚寒", "虚": "太阴-里虚寒", "平": "太阴-里虚寒"},
+            "少阴": {"实": "少阴-里虚寒", "虚": "少阴-里虚寒", "平": "少阴-里虚寒"},
+            "厥阴": {"实": "厥阴-上热下寒", "虚": "厥阴-血虚寒厥", "平": "厥阴-上热下寒"}
+        }
+        
+        目标证型 = 映射[六经][证型]
+        匹配方名 = 伤寒杂病论类.根据证型查方名法(目标证型)
+        
+        if 匹配方名:
+            方 = 伤寒杂病论类.方剂库[匹配方名]
+            return f"【{匹配方名}】{'、'.join(方['组成'])}"
+        return ""
+    
+    # 五运六气病机伤寒选方
+    @staticmethod
+    def 五运六气病机伤寒选方(病机: str) -> str:
+        """五运六气病机自动推方"""
+        配置表 = [
+            ("肝", ["风动","不舒","亢盛","火旺"], "厥阴-上热下寒"),
+            ("肝", ["不足","虚寒","血亏"], "厥阴-血虚寒厥"),
+            ("心", ["火旺","偏亢","不宁","上炎"], "少阴-里热实"), # 假设库里有此证型
+            ("心", ["不足","虚烦","阳虚"], "少阴-里虚寒"),
+            ("脾", ["湿困","壅滞","实满"], "太阴-里虚寒"), # 太阴以温中为主
+            ("肺", ["燥伤","气逆","痰盛"], "阳明-里热盛"),
+            ("肾", ["寒凝","水停","热结"], "少阴-里虚寒"),
+        ]
+        for 脏, 关键词, 对应证型 in 配置表:
+            if 脏 in 病机 and any(k in 病机 for k in 关键词):
+                匹配方名 = 伤寒杂病论类.根据证型查方名法(对应证型)
+                if 匹配方名:
+                    方 = 伤寒杂病论类.方剂库[匹配方名]
+                    return f"【{匹配方名}】{'、'.join(方['组成'])}"
