+++ custom_components/tianyuan_calendar/tianyuan/qihuang/wuyun_liuqi.py
@@ -1,0 +1,468 @@
+"""天元岐黄引擎 - 五运六气."""
+from __future__ import annotations
+
+from ..library import 加载全量映射表类
+from .fuxingjue import 辅行诀脏腑用药法要类
+from .shanghan import 伤寒杂病论类
+
+maps = 加载全量映射表类()
+五运六气映射 = maps.get("五运六气类", {})
+
+class 五运六气类:
+    """五运六气全功能推演类"""
+
+    五运表 = 五运六气映射.get("五运表", {})
+    六气表 = 五运六气映射.get("六气表", {})                     # 年支 -> 司天
+    五运偏盛偏衰 = 五运六气映射.get("五运偏盛偏衰", {})
+    六气偏盛偏衰 = 五运六气映射.get("六气偏盛偏衰", {})
+    六步主气表 = 五运六气映射.get("六步主气", [])               # 步名列表
+    六气轮转表 = 五运六气映射.get("六气轮转", [])               # 三阴三阳固定次序
+    六气属性表 = 五运六气映射.get("六气属性", {})
+    五行关系表 = 五运六气映射.get("五行关系", {})
+    气机病机表 = 五运六气映射.get("气机病机", {})
+    主气六步序列表 = 五运六气映射.get("主气六步序列", [])
+    司天在泉对表 = 五运六气映射.get("司天在泉对", {})
+    年支五行表 = 五运六气映射.get("年支五行", {})
+    三阴三阳顺序表 = 五运六气映射.get("三阴三阳顺序", [])
+    标本中气表 = 五运六气映射.get("标本中气", [])
+
+    # ========= 全局缓存 =========
+    _年度缓存 = {}   # 键："年干+年支" -> 年度数据
+
+    # ========= 基础取数（无缓存，高频调用） =========
+    @staticmethod
+    def 获取五运法(年干: str) -> str:
+        return 五运六气类.五运表.get(年干, "未知")
+
+    @staticmethod
+    def 获取六气法(年支: str) -> str:          # 返回司天
+        return 五运六气类.六气表.get(年支, "未知")
+
+    @staticmethod
+    def 获取在泉法(年支: str) -> str:
+        司天 = 五运六气类.获取六气法(年支)
+        return 五运六气类.司天在泉对表.get(司天, "未知")
+
+    @staticmethod
+    def 获取五运偏差法(五运: str) -> list:
+        return 五运六气类.五运偏盛偏衰.get(五运, ["未知", "未知"])
+
+    @staticmethod
+    def 获取六气偏差法(六气: str) -> list:
+        return 五运六气类.六气偏盛偏衰.get(六气, ["未知", "未知"])
+
+    # ========= 六步主气（固定） =========
+    @staticmethod
+    def 获取六步主气法() -> list:
+        return list(zip(五运六气类.六步主气表, 五运六气类.主气六步序列表))
+
+    # ========= 六步客气（经典司天在泉法） =========
+    @staticmethod
+    def 获取六步客气法(年支: str) -> list:
+        """
+        按年支定司天（三之气）→ 在泉（终之气），
+        依据三阴三阳固定次序推算六步客气。
+        """
+        司天 = 五运六气类.获取六气法(年支)
+        在泉 = 五运六气类.获取在泉法(年支)
+        顺序 = 五运六气类.三阴三阳顺序表
+
+        if 司天 not in 顺序 or 在泉 not in 顺序:
+            return [(步, "未知") for 步 in 五运六气类.六步主气表]
+
+        # 以三之气为司天，逆推初之气、二之气
+        司天位置 = 顺序.index(司天)
+        # 四之气：司天后一位
+        四气位置 = (司天位置 + 1) % 6
+        # 五之气：司天后两位
+        五气位置 = (司天位置 + 2) % 6
+        # 终之气（在泉）应位于司天后三位（对化关系），校验一下
+        终气位置 = (司天位置 + 3) % 6
+        if 顺序[终气位置] != 在泉:
+            # 数据异常时回退为直接使用轮转表
+            终气位置 = 顺序.index(在泉)
+        # 初之气：在泉后一位
+        初气位置 = (终气位置 + 1) % 6
+        # 二之气：在泉后两位
+        二气位置 = (终气位置 + 2) % 6
+
+        # 六步客气序列：初、二、三、四、五、终
+        客气序列 = [
+            顺序[初气位置],
+            顺序[二气位置],
+            顺序[司天位置],
+            顺序[四气位置],
+            顺序[五气位置],
+            顺序[终气位置]
+        ]
+        return list(zip(五运六气类.六步主气表, 客气序列))
+
+    # ========= 气机 -> 病机（基础） =========
+    @staticmethod
+    def 获取病机法(气机: str) -> dict:
+        return 五运六气类.气机病机表.get(气机, {
+            "性质": "中和", "病机": "气机调畅", "症候": "平", "脏腑": "无", "经络": "无"
+        })
+
+    # ========= 运气同化计算 =========
+    @staticmethod
+    def 计算运气同化法(年干: str, 年支: str) -> dict:
+        """返回天符、岁会等特殊年份判定"""
+        岁运 = 五运六气类.获取五运法(年干)       # 如 "木运"
+        岁运五行 = 岁运.replace("运", "")        # "木"
+        司天 = 五运六气类.获取六气法(年支)       # "厥阴风木"
+        在泉 = 五运六气类.获取在泉法(年支)
+        司天五行 = 五运六气类.六气属性表[司天]["五行"]
+        在泉五行 = 五运六气类.六气属性表[在泉]["五行"]
+        年支五行值 = 五运六气类.年支五行表[年支]
+
+        # 天符：岁运五行 == 司天五行
+        天符 = (岁运五行 == 司天五行)
+        # 岁会：岁运五行 == 年支五行
+        岁会 = (岁运五行 == 年支五行值)
+        # 同天符：阳年（甲丙戊庚壬）且岁运太过，岁运五行 == 在泉五行
+        阳干 = set("甲丙戊庚壬")
+        同天符 = (年干 in 阳干) and (岁运五行 == 在泉五行)
+        # 同岁会：阴年（乙丁己辛癸）且岁运不及，岁运五行 == 在泉五行
+        阴干 = set("乙丁己辛癸")
+        同岁会 = (年干 in 阴干) and (岁运五行 == 在泉五行)
+        # 太乙天符：天符 + 岁会
+        太乙天符 = 天符 and 岁会
+
+        描述列表 = []
+        if 天符:
+            描述列表.append("天符")
+        if 岁会:
+            描述列表.append("岁会")
+        if 同天符:
+            描述列表.append("同天符")
+        if 同岁会:
+            描述列表.append("同岁会")
+        if 太乙天符:
+            描述列表.append("太乙天符")
+        描述 = "，".join(描述列表) if 描述列表 else "平气"
+
+        return {
+            "天符": 天符, "岁会": 岁会,
+            "同天符": 同天符, "同岁会": 同岁会,
+            "太乙天符": 太乙天符,
+            "描述": 描述
+        }
+
+    # ========= 胜复推导 =========
+    @staticmethod
+    def 计算胜复法(五运: str, 司天: str) -> dict:
+        """
+        根据岁运太过与司天，给出可能的胜气、复气及影响。
+        简化模型：五运太过则本气胜，所不胜之子来复。
+        """
+        # 五行相胜关系：木胜土，金胜木，火胜金，水胜火，土胜水
+        胜克 = {"木": "土", "火": "金", "土": "水", "金": "木", "水": "火"}
+        # 五行相生关系（子来复）：如木胜土，土之子金复木
+        子复 = {"土": "金", "金": "水", "水": "木", "木": "火", "火": "土"}
+
+        # 判断岁运是否太过
+        太过不及 = 五运六气类.获取五运偏差法(五运)
+        是太过 = "太过" in 太过不及[0]
+
+        if not 是太过:
+            return {"胜气": "无（岁运不及）", "复气": "无", "病机提示": "胜复不显"}
+
+        岁运五行 = 五运.replace("运", "")
+        所胜 = 胜克.get(岁运五行)
+        复气五行 = 子复.get(所胜) if 所胜 else None
+
+        司天气机 = 五运六气类.六气属性表[司天]["五行"] if 司天 in 五运六气类.六气属性表 else None
+
+        # 如果司天五行与岁运五行同，则胜气更烈
+        加临 = "司天同化，胜气倍增" if 司天气机 == 岁运五行 else "司天不同化"
+
+        复气描述 = f"{复气五行}气来复" if 复气五行 else "复气不明"
+        return {
+            "胜气": f"{岁运五行}气偏胜",
+            "所胜": f"{所胜}气受抑",
+            "复气": f"{复气五行}气来复（{所胜}之子）" if 复气五行 else "待考",
+            "病机提示": f"{岁运五行}胜乘{所胜}，{加临}，{复气描述}。",
+        }
+
+    # ========= 客主相交（升级版：含标本中气） =========
+    @staticmethod
+    def 客主相交法(客气: str, 主气: str) -> dict:
+        # 同气相求
+        if 客气 == 主气:
+            气机 = 五运六气类.六气属性表[客气]["气机"]
+            病机 = 五运六气类.获取病机法(气机)
+            标本 = 五运六气类.标本中气表[客气]
+        
+            return {
+                "关系": "同气相求",
+                "五行关系": "同类相应",
+                "病机": 病机["病机"],
+                "症候": 病机["症候"],
+                "脏腑": 病机["脏腑"],
+                "经络": 病机["经络"],
+                "标本从化": 标本["从化"],
+                "摘要": f"{客气} 同气相求 → {病机['病机']}"
+            }
+
+        客五行 = 五运六气类.六气属性表[客气]["五行"]
+        主五行 = 五运六气类.六气属性表[主气]["五行"]
+        客标本 = 五运六气类.标本中气表[客气]
+        主标本 = 五运六气类.标本中气表[主气]
+
+        # 君臣顺逆：主为君，客为臣。
+        生克关系 = None
+        原始方向 = None
+        正向键 = f"{客五行},{主五行}"
+        反向键 = f"{主五行},{客五行}"
+        if 正向键 in 五运六气类.五行关系表:
+            生克关系 = 五运六气类.五行关系表[正向键]
+            原始方向 = "客→主"
+        elif 反向键 in 五运六气类.五行关系表:
+            生克关系 = 五运六气类.五行关系表[反向键]
+            原始方向 = "主→客"
+            if "生" in 生克关系:
+                生克关系 = 生克关系.replace("生", "被生")
+            elif "克" in 生克关系:
+                生克关系 = 生克关系.replace("克", "被克")
+
+        # 相得/不相得判定
+        if not 生克关系:
+            顺逆 = "未知"
+        elif "生" in 生克关系:
+            if 原始方向 == "客→主":
+                顺逆 = "相得（客气生主气）"
+            else:
+                顺逆 = "不相得（主气生客气）"
+        else:  # 克
+            if 原始方向 == "客→主":
+                顺逆 = "不相得（客气克主气，逆）"
+            else:
+                顺逆 = "相得（主气克客气，君胜臣）"
+
+        # 偏盛乘侮
+        客太过, _ = 五运六气类.获取六气偏差法(客气)
+        主太过, _ = 五运六气类.获取六气偏差法(主气)
+        客偏盛 = "太过" in 客太过
+        主偏盛 = "太过" in 主太过
+
+        乘侮 = ""
+        if 客偏盛 and "克" in (生克关系 or ""):
+            if "被克" in 生克关系:   # 主克客，但客偏盛反侮
+                乘侮 = "侮"
+            else:                     # 客克主，乘
+                乘侮 = "乘"
+        elif 主偏盛 and "克" in (生克关系 or ""):
+            if "被克" in 生克关系:   # 客克主，主偏盛反侮
+                乘侮 = "侮"
+            else:                     # 主克客，乘
+                乘侮 = "乘"
+
+        # 主导气机
+        if 客偏盛 and not 主偏盛:
+            主导气机 = 五运六气类.六气属性表[客气]["气机"]
+            主导类型 = "客气偏盛"
+        elif 主偏盛 and not 客偏盛:
+            主导气机 = 五运六气类.六气属性表[主气]["气机"]
+            主导类型 = "主气偏盛"
+        elif 客偏盛 and 主偏盛:
+            # 两盛则看生克：客胜主为从，较顺；主胜客为逆
+            if "被克" in (生克关系 or ""):
+                主导气机 = 五运六气类.六气属性表[客气]["气机"]
+                主导类型 = "客主俱盛，客气反胜（从）"
+            else:
+                主导气机 = 五运六气类.六气属性表[主气]["气机"]
+                主导类型 = "客主俱盛，主气胜客（逆）"
+        else:
+            # 皆不偏盛，按常规胜气
+            if "被克" in (生克关系 or ""):
+                主导气机 = 五运六气类.六气属性表[客气]["气机"]
+                主导类型 = "客气胜主（从）"
+            else:
+                主导气机 = 五运六气类.六气属性表[主气]["气机"]
+                主导类型 = "主气胜客（逆）"
+
+        病机 = 五运六气类.获取病机法(主导气机)
+        五行描述 = f"{生克关系 or '平气'}{'（' + 乘侮 + '）' if 乘侮 else ''}"
+
+        # 标本从化提示
+        标本提示 = f"客从{客标本['从化']}，主从{主标本['从化']}"
+
+        return {
+            "关系": "客主相交",
+            "五行关系": 五行描述,
+            "顺逆": 顺逆,
+            "主导类型": 主导类型,
+            "主导气机": 主导气机,
+            "病机": 病机["病机"],
+            "症候": 病机["症候"],
+            "脏腑": 病机["脏腑"],
+            "经络": 病机["经络"],
+            "标本从化": 标本提示,
+            "摘要": f"{客气} vs {主气}：{五行描述} {顺逆} {主导类型} → {病机['病机']}"
+        }
+
+    # ========= 六步运气数据 =========
+    @staticmethod
+    def 获取当前步序法(农历) -> int:
+        """判定当前处于六气中的第几步 (0-5)"""
+        节点 = ["大寒", "春分", "小满", "大暑", "秋分", "小雪"]
+        当前阳历 = 农历.getSolar().toYmd()
+        节气表 = 农历.getJieQiTable()
+        # 从最后一步往回查
+        for 序号 in range(5, -1, -1):
+            目标 = 节气表.get(节点[序号])
+            if 目标 and 当前阳历 >= 目标.toYmd():
+                return 序号
+        return 5  # 兜底
+
+    @staticmethod
+    def 六步运气数据类法(年干: str, 年支: str) -> list:
+        """
+        全量排盘计算：生成全年 6 个步序的详细数据列表
+        """
+        主气序列 = 五运六气类.获取六步主气法()
+        客气序列 = 五运六气类.获取六步客气法(年支) # 遵循要求：改用年支计算客气起点
+
+        结果 = []
+        for (步, 主气), (_, 客气) in zip(主气序列, 客气序列):
+            关系 = 五运六气类.客主相交法(客气, 主气)
+            结果.append({
+                "state": 步,
+                "attributes": {
+                    "主气": 主气,
+                    "客气": 客气,
+                    "五行关系": 关系["五行关系"],
+                    "顺逆": 关系.get("顺逆", ""),
+                    "主导类型": 关系.get("主导类型", ""),
+                    "病机": 关系["病机"],
+                    "症候": 关系["症候"],
+                    "脏腑": 关系["脏腑"],
+                    "经络": 关系["经络"],
+                    "标本从化": 关系["标本从化"],
+                    # 跨类联动：获取对应的方剂参考
+                    "辅行诀参考": 辅行诀脏腑用药法要类.五运六气病机辅行诀选方(关系["病机"]),
+                    "经方参考": 伤寒杂病论类.五运六气病机伤寒选方(关系["病机"]),
+                    "摘要": f"{步}：{关系['摘要']}"
+                }
+            })
+        return 结果
+
+    # ========= 年度五运 =========
+    @staticmethod
+    def 年五运类法(年干: str) -> dict:
+        五运 = 五运六气类.获取五运法(年干)
+        太过, 不及 = 五运六气类.获取五运偏差法(五运)
+        摘要 = f"{五运}（{太过} / {不及}）"
+        
+        return {
+            "state": f"{五运}{'太过' if '太过' in 太过 else '不及'}",
+            "attributes": {
+                "五运": 五运, 
+                "太过": 太过, 
+                "不及": 不及,
+                "摘要": 摘要
+            }
+        }
+
+    # ========= 年度六气 =========
+    @staticmethod
+    def 年六气类法(年支: str) -> dict:
+        司天 = 五运六气类.获取六气法(年支)
+        在泉 = 五运六气类.获取在泉法(年支)
+        太过, 不及 = 五运六气类.获取六气偏差法(司天)
+        摘要 = f"{司天}（{太过} / {不及}）"
+
+        return {
+            "state": f"{司天}司天·{在泉}在泉",
+            "attributes": {
+                "司天": 司天, 
+                "在泉": 在泉, 
+                "司天太过": 太过, 
+                "司天不及": 不及,
+                "摘要": 摘要
+                }
+        }
+
+    # ========= 年度总览（含运气同化、胜复） =========
+    @staticmethod
+    def 年度总览类(年干: str, 年支: str) -> dict:
+        五运 = 五运六气类.年五运类法(年干)
+        六气 = 五运六气类.年六气类法(年支)
+        六步 = 五运六气类.六步运气数据类法(年干, 年支)
+        同化 = 五运六气类.计算运气同化法(年干, 年支)
+        胜复 = 五运六气类.计算胜复法(五运["attributes"]["五运"], 六气["attributes"]["司天"])
+
+        扁平属性 = {}
+        # 平铺五运
+        扁平属性["岁运"] = 五运["attributes"]["五运"]
+        扁平属性["岁运太过"] = 五运["attributes"]["太过"]
+        扁平属性["岁运不及"] = 五运["attributes"]["不及"]
+        扁平属性["岁运摘要"] = 五运["attributes"]["摘要"]
+        # 平铺六气
+        扁平属性["司天"] = 六气["attributes"]["司天"]
+        扁平属性["在泉"] = 六气["attributes"]["在泉"]
+        扁平属性["司天太过"] = 六气["attributes"]["司天太过"]
+        扁平属性["司天不及"] = 六气["attributes"]["司天不及"]
+        扁平属性["司天摘要"] = 六气["attributes"]["摘要"]
+        # 平铺同化
+        扁平属性["运气同化类型"] = 同化["描述"]
+        扁平属性["是否为天符"] = 同化["天符"]
+        扁平属性["是否为岁会"] = 同化["岁会"]
+        扁平属性["是否为同天符"] = 同化["同天符"]
+        扁平属性["是否为同岁会"] = 同化["同岁会"]
+        扁平属性["是否为太乙天符"] = 同化["太乙天符"]
+        # 平铺胜复
+        扁平属性["胜气"] = 胜复.get("胜气", "")
+        扁平属性["所胜受抑"] = 胜复.get("所胜", "")
+        扁平属性["复气"] = 胜复.get("复气", "")
+        扁平属性["胜复病机提示"] = 胜复.get("病机提示", "")
+        # 平铺六步摘要（每步一行）
+        六步摘要列表 = []
+        for 步数据 in 六步:
+            步名 = 步数据["state"]
+            主气 = 步数据["attributes"]["主气"]
+            客气 = 步数据["attributes"]["客气"]
+            病机 = 步数据["attributes"]["病机"]
+            六步摘要列表.append(f"{步名}：主气{主气}，客气{客气}，病机{病机}")
+        扁平属性["六步摘要"] = "；".join(六步摘要列表)
+        # 添加年度摘要
+        扁平属性["年度摘要"] = f"{年干}{年支}年：{五运['state']}，{六气['state']}。{同化['描述']}。{胜复['病机提示']}"
+
+        return {
+            "state": f"{年干}{年支}年运气总览",
+            "attributes": 扁平属性
+        }
+
+    # ========= 全量计算（带缓存） =========
+    @staticmethod
+    def 全量计算类(农历) -> dict:
+        年干 = 农历.getYearGanExact()
+        年支 = 农历.getYearZhiExact()
+        缓存键 = f"{年干}{年支}"
+
+        if 缓存键 in 五运六气类._年度缓存:
+            年度数据 = 五运六气类._年度缓存[缓存键]
+        else:
+            年度数据 = {
+                "年五运数据": 五运六气类.年五运类法(年干),
+                "年六气数据": 五运六气类.年六气类法(年支),
+                "六步运气数据_全量": 五运六气类.六步运气数据类法(年干, 年支),
+                "年度总览数据": 五运六气类.年度总览类(年干, 年支)
+            }
+            五运六气类._年度缓存[缓存键] = 年度数据
+
+        # 取出当前步
+        全量六步 = 五运六气类.六步运气数据类法(年干, 年支)
+        当前索引 = 五运六气类.获取当前步序法(农历)
+        当前步 = 全量六步[当前索引]
+
+        return {
+            "年五运数据": 年度数据["年五运数据"],
+            "年六气数据": 年度数据["年六气数据"],
+            "六步运气数据": {
+                "state": 当前步["state"],
+                "attributes": 当前步["attributes"]
+            },
+            "年度总览数据": 年度数据["年度总览数据"]
+        }
