+++ custom_components/tianyuan_calendar/tianyuan/qihuang/ziwuliu.py
@@ -1,0 +1,442 @@
+"""天元岐黄引擎 - 子午流注（纳甲 / 纳子 / 灵龟八法 / 飞腾八法 / 迎随补泻）."""
+from __future__ import annotations
+
+from ..library import 加载全量映射表类
+from .fuxingjue import 辅行诀脏腑用药法要类
+from .shanghan import 伤寒杂病论类
+
+maps = 加载全量映射表类()
+子午流注映射 = maps.get("子午流注类", {})
+
+class 子午流注类:
+    """子午流注：纳甲法 / 纳子法 / 灵龟八法 / 飞腾八法 / 迎随补泻"""
+
+    穴位手册 = 子午流注映射.get("穴位手册", {})
+    时辰经络 = 子午流注映射.get("时辰经络", [])
+    流注顺序 = 子午流注映射.get("十二经流注顺序", [])
+    子午相对 = 子午流注映射.get("子午相对", [])
+    纳甲取穴 = 子午流注映射.get("纳甲取穴", [])
+    值日经 = 子午流注映射.get("六十甲子值日经", [])
+    八卦方位 = 子午流注映射.get("八卦方位", {})
+    八卦吉凶 = 子午流注映射.get("八卦吉凶", {})
+    飞腾起点 = 子午流注映射.get("飞腾起点", "")
+    五输穴表 = 子午流注映射.get("五输穴表", {})
+    纳子补泻表 = 子午流注映射.get("纳子补泻表", {})
+    飞腾配穴表 = 子午流注映射.get("飞腾配穴表", {})
+
+    # ========= 公共工具 =========
+    @staticmethod
+    def 小时转时支(小时: int) -> str:
+        对照 = [
+            "子","丑","丑","寅","寅","卯","卯","辰","辰","巳","巳","午",
+            "午","未","未","申","申","酉","酉","戌","戌","亥","亥","子"
+        ]
+        return 对照[小时]
+
+    @staticmethod
+    def 穴位详情(穴位: str) -> dict:
+        return 子午流注类.穴位手册.get(穴位, {"位置": "未知", "功能": "未知", "针法": "未知"})
+
+    @staticmethod
+    def 生成描述(农历) -> str:
+        日干支 = 农历.getDayInGanZhiExact2()
+        时干支 = 农历.getTimeInGanZhi()
+        时支 = 农历.getTimeZhi()
+        return f"{日干支}日{时干支}时（{时支}时）"
+
+    @staticmethod
+    def 顺经(经络: str) -> str:
+        seq = 子午流注类.流注顺序
+        if 经络 not in seq:
+            return "未知"
+        return seq[(seq.index(经络) + 1) % len(seq)]
+
+    @staticmethod
+    def 逆经(经络: str) -> str:
+        seq = 子午流注类.流注顺序
+        if 经络 not in seq:
+            return "未知"
+        return seq[(seq.index(经络) - 1) % len(seq)]
+
+    @staticmethod
+    def 八卦信息(卦: str) -> tuple:
+        return (
+            子午流注类.八卦方位.get(卦, "未知"),
+            子午流注类.八卦吉凶.get(卦, "未知"),
+        )
+
+    @staticmethod
+    def 判开阖枢(时支: str) -> dict:
+        开 = ["子", "午", "卯", "酉"]
+        阖 = ["丑", "辰", "未", "戌"]
+        枢 = ["寅", "巳", "申", "亥"]
+
+        if 时支 in 开:
+            return {"法": "开", "说明": "开法：气血初开，多取井穴、荥穴，以开其端。"}
+        if 时支 in 阖:
+            return {"法": "阖", "说明": "阖法：气血闭藏，多取输穴、合穴，以收其气。"}
+        if 时支 in 枢:
+            return {"法": "枢", "说明": "枢法：气血转折，多取经穴、合穴，以调其枢机。"}
+        return {"法": "未知", "说明": "未能判定开/阖/枢。"}
+    
+    # ========= 纳甲法=========
+    @staticmethod
+    def 纳甲法类(农历):
+        """
+        严格按徐文伯纳甲法：阳日阳时开阳经，阴日阴时开阴经。
+        井荥输经合逐时顺移，闭穴时用合日互用（甲与己合等）。
+        返回当前时辰应开的穴位（含五行属性自动推算）。
+        """
+        日干 = 农历.getDayGanExact2()
+        时干 = 农历.getTimeGan()
+        时支 = 农历.getTimeZhi()
+        日干支 = 农历.getDayInGanZhiExact2()
+        当前时干支 = f"{时干}{时支}"
+
+        # 合日互用配对
+        合日 = {"甲":"己","己":"甲","乙":"庚","庚":"乙","丙":"辛","辛":"丙",
+                "丁":"壬","壬":"丁","戊":"癸","癸":"戊"}
+
+        # 阳经/阴经顺序（纳甲经典顺序）
+        阳经序 = ["胆经","小肠经","胃经","大肠经","膀胱经","三焦经"]
+        阴经序 = ["肝经","心经","脾经","肺经","肾经","心包经"]
+
+        # 开井时辰表（日干 -> (开井时辰干支, 经络, 井穴)）
+        开井表 = {
+            "甲": ("甲戌", "胆经", "足窍阴"),
+            "乙": ("乙酉", "肝经", "大敦"),
+            "丙": ("丙申", "小肠经", "少泽"),
+            "丁": ("丁未", "心经", "少冲"),
+            "戊": ("戊午", "胃经", "厉兑"),
+            "己": ("己巳", "脾经", "隐白"),
+            "庚": ("庚辰", "大肠经", "商阳"),
+            "辛": ("辛卯", "肺经", "少商"),
+            "壬": ("壬寅", "膀胱经", "至阴"),
+            "癸": ("癸亥", "肾经", "涌泉")
+        }
+
+        # 五输穴的五行属性（井荥输经合）
+        # 阳经：井金 荥水 输木 经火 合土
+        # 阴经：井木 荥火 输土 经金 合水
+        阳经五输五行 = {"井":"金", "荥":"水", "输":"木", "经":"火", "合":"土"}
+        阴经五输五行 = {"井":"木", "荥":"火", "输":"土", "经":"金", "合":"水"}
+
+        # 工具函数：从开井时辰开始，遍历12个时辰，查找当前时辰对应的穴位
+        def 从开井时辰查找(开井干支, 经络, 井穴):
+            开井干 = 开井干支[0]
+            开井支 = 开井干支[1]
+            天干序 = ["甲","乙","丙","丁","戊","己","庚","辛","壬","癸"]
+            地支序 = ["子","丑","寅","卯","辰","巳","午","未","申","酉","戌","亥"]
+            天干始 = 天干序.index(开井干)
+            地支始 = 地支序.index(开井支)
+
+            # 确定经序
+            if 经络 in 阳经序:
+                经序 = 阳经序
+            else:
+                经序 = 阴经序
+            经索引 = 经序.index(经络)
+
+            输穴列表 = ["井","荥","输","经","合"]
+            穴序 = 0   # 从井穴开始
+
+            for 偏移 in range(12):
+                干 = 天干序[(天干始 + 偏移) % 10]
+                支 = 地支序[(地支始 + 偏移) % 12]
+                干支 = f"{干}{支}"
+                if 干支 == 当前时干支:
+                    当前经 = 经序[经索引]
+                    输穴类型 = 输穴列表[穴序]
+                    穴位 = 子午流注类.五输穴表[当前经][输穴类型]
+                    五行 = (阳经五输五行 if 当前经 in 阳经序 else 阴经五输五行)[输穴类型]
+                    return 穴位, 输穴类型, 当前经, 五行
+                # 移到下一个穴或经
+                穴序 += 1
+                if 穴序 >= 5:   # 井荥输经合走完
+                    穴序 = 0
+                    经索引 = (经索引 + 1) % len(经序)
+            return None, None, None, None
+
+        # 先试本日开井
+        穴位结果 = None
+        if 日干 in 开井表:
+            开井干支, 经络, 井穴 = 开井表[日干]
+            穴位结果, 输穴类型, 使用经络, 五行属性 = 从开井时辰查找(开井干支, 经络, 井穴)
+
+        # 若未找到，试合日开井
+        if not 穴位结果:
+            合日干 = 合日.get(日干)
+            if 合日干 and 合日干 in 开井表:
+                开井干支, 经络, 井穴 = 开井表[合日干]
+                穴位结果, 输穴类型, 使用经络, 五行属性 = 从开井时辰查找(开井干支, 经络, 井穴)
+
+        # 组装返回数据
+        描述 = 子午流注类.生成描述(农历)
+        值日经 = 子午流注类.值日经.get(日干支, "未知")
+        三法 = 子午流注类.判开阖枢(时支)
+
+        if 穴位结果:
+            穴位详情 = 子午流注类.穴位详情(穴位结果)
+            摘要 = f"{穴位结果}（{输穴类型}，{使用经络}，五行{五行属性}），采用{三法['法']}法"
+            return {
+                "state": 穴位结果,
+                "attributes": {
+                    "值日经": 值日经,
+                    "穴位": 穴位结果,
+                    "输穴": 输穴类型,
+                    "经络": 使用经络,
+                    "五行": 五行属性,
+                    "三法": 三法["法"],
+                    "三法说明": 三法["说明"],
+                    "位置": 穴位详情['位置'],
+                    "功能": 穴位详情['功能'],
+                    "针法": 穴位详情['针法'],
+                    "摘要": 摘要,
+                    "描述": 描述,
+                    "详细描述": (
+                        f"{描述}，值日经：{值日经}，采用{三法['法']}法，"
+                        f"取{使用经络}之{输穴类型}——{穴位结果}（五行{五行属性}）。\n"
+                        f"穴位位置：{穴位详情['位置']}\n"
+                        f"主要功能：{穴位详情['功能']}\n"
+                        f"针刺方法：{穴位详情['针法']}\n"
+                        f"三法说明：{三法['说明']}"
+                    )
+                }
+            }
+        else:
+            # 闭穴，但尽量给出合日穴位的提示（如果合日仍闭穴，则真的闭穴）
+            提示 = "闭穴，请用合日互用或纳子法"
+            合日干 = 合日.get(日干)
+            if 合日干 and 合日干 in 开井表:
+                提示 += f"（可尝试{合日干}日开穴）"
+            return {
+                "state": "闭穴",
+                "attributes": {
+                    "描述": 描述,
+                    "详细描述": f"{描述}，当前时辰无纳甲开穴。{提示}"
+                }
+            }
+
+    # ========= 纳子法（补泻增强版） =========
+    @staticmethod
+    def 纳子法类(dt, 证型: str = "平"):
+        """
+        证型：平（取本穴/原穴）、实（取子穴）、虚（取母穴，下时辰取）
+        """
+        小时 = dt.hour
+        时支 = 子午流注类.小时转时支(小时)
+        经络, 默认穴, 默认输穴, 时间段 = 子午流注类.时辰经络[时支]
+        补泻信息 = 子午流注类.纳子补泻表.get(时支, {})
+        穴位详情 = {}
+        穴位 = 默认穴
+        输穴类型 = 默认输穴
+
+        if 证型 == "实":
+            实证穴 = 补泻信息.get("实证穴", 默认穴)
+            穴位 = 实证穴
+            输穴类型 = "子穴"
+            补泻描述 = f"实证泻子：取{经络}子穴{穴位}"
+        elif 证型 == "虚":
+            虚证穴 = 补泻信息.get("虚证穴", 默认穴)
+            虚证时辰 = 补泻信息.get("虚证时辰", 时支)
+            穴位 = 虚证穴
+            输穴类型 = "母穴"
+            # 提示应在下一个时辰取穴，但为了方便，这里仍显示穴位，并在描述中提示
+            补泻描述 = f"虚证补母：宜在{虚证时辰}时取{经络}母穴{穴位}"
+        else:
+            补泻描述 = f"平补平泻：取本穴/原穴{穴位}"
+
+        穴位详情 = 子午流注类.穴位详情(穴位)
+        对应时支 = 子午流注类.子午相对.get(时支)
+        对应经络 = 子午流注类.时辰经络[对应时支][0] if 对应时支 else "未知"
+
+        辅行诀参考 = 辅行诀脏腑用药法要类.子午流注纳子法辅行诀选方(经络, 证型)
+        伤寒杂病论参考 = 伤寒杂病论类.子午流注纳子法伤寒选方(经络, 证型)
+
+        描述 = f"{时支}时（{时间段}）"
+        摘要 = f"{时支}时当令{经络}，{补泻描述}，子午相对：{对应经络}"
+
+        return {
+            "state": 经络,
+            "attributes": {
+                "时支": 时支,
+                "时间段": 时间段,
+                "经络": 经络,
+                "穴位": 穴位,
+                "输穴": 输穴类型,
+                "子午相对时支": 对应时支,
+                "子午相对经络": 对应经络,
+                "位置": 穴位详情['位置'],
+                "功能": 穴位详情['功能'],
+                "针法": 穴位详情['针法'],
+                "摘要": 摘要,
+                "描述": 描述,
+                "辅行诀参考": 辅行诀参考,
+                "经方参考": 伤寒杂病论参考,
+                "详细描述": (
+                    f"{描述}，气血流注于{经络}。\n"
+                    f"{补泻描述}。\n"
+                    f"子午相对：{时支} ↔ {对应时支}，经络：{经络} ↔ {对应经络}。\n"
+                    f"穴位位置：{穴位详情['位置']}\n"
+                    f"主要功能：{穴位详情['功能']}\n"
+                    f"针刺方法：{穴位详情['针法']}"
+                )
+            }
+        }
+
+    # ========= 灵龟八法 =========
+    @staticmethod
+    def 灵龟八法类(农历, 性别: str):
+        """子午流注类 · 灵龟八法"""
+
+        天干数 = {'甲':9,'己':9,'乙':8,'庚':8,'丙':7,'辛':7,'丁':6,'壬':6,'戊':5,'癸':5}
+        地支数 = {'子':9,'午':9,'丑':8,'未':8,'寅':7,'申':7,'卯':6,'酉':6,'辰':5,'戌':5,'巳':4,'亥':4}
+
+        日干 = 农历.getDayGanExact2()
+        日支 = 农历.getDayZhiExact2()
+        时干 = 农历.getTimeGan()
+        时支 = 农历.getTimeZhi()
+
+        总和 = 天干数[日干] + 地支数[日支] + 天干数[时干] + 地支数[时支]
+
+        初数 = 9 if 日干 in ['甲','丙','戊','庚','壬'] else 6
+        余数 = 总和 % 初数 or 初数
+
+        if 余数 == 5:
+            余数 = 2 if 性别 == "男" else 8
+
+        八法表 = {
+            1:('申脉','阳跷脉','坎','坎一宫'),
+            2:('照海','阴跷脉','坤','坤二宫'),
+            3:('外关','阳维脉','震','震三宫'),
+            4:('足临泣','带脉','巽','巽四宫'),
+            6:('公孙','冲脉','乾','乾六宫'),
+            7:('后溪','督脉','兑','兑七宫'),
+            8:('内关','阴维脉','艮','艮八宫'),
+            9:('列缺','任脉','离','离九宫')
+        }
+
+        穴位, 经络, 卦, 宫位 = 八法表.get(余数)
+        穴位详情 = 子午流注类.穴位详情(穴位)
+        方位, 吉凶 = 子午流注类.八卦信息(卦)
+
+        描述 = 子午流注类.生成描述(农历)
+        摘要 = f"{穴位}（{经络}），属{卦}卦（{方位}），{吉凶}"
+
+        return {
+            "state": 穴位,
+            "attributes": {
+                "数值": 余数,
+                "卦": 卦,
+                "宫位": 宫位,
+                "方位": 方位,
+                "吉凶": 吉凶,
+                "穴位": 穴位,
+                "经络": 经络,
+                "位置": 穴位详情['位置'],
+                "功能": 穴位详情['功能'],
+                "针法": 穴位详情['针法'],
+                "摘要": 摘要,
+                "描述": 描述,
+                "详细描述": (
+                    f"{描述}，灵龟八法得数：{余数}，对应{宫位}（{卦}卦，方位：{方位}，吉凶：{吉凶}）。\n"
+                    f"取穴：{穴位}（{经络}）。\n"
+                    f"穴位位置：{穴位详情['位置']}\n"
+                    f"主要功能：{穴位详情['功能']}\n"
+                    f"针刺方法：{穴位详情['针法']}"
+                )
+            }
+        }
+
+    # ========= 飞腾八法（补配穴） =========
+    @staticmethod
+    def 飞腾八法类(农历):
+        """子午流注类 · 飞腾八法"""
+
+        日干 = 农历.getDayGanExact2()
+        时支 = 农历.getTimeZhi()
+
+        起点 = 子午流注类.飞腾起点.get(日干, "子")
+        时序 = ["子","丑","寅","卯","辰","巳","午","未","申","酉","戌","亥"]
+
+        飞腾数 = (时序.index(时支) - 时序.index(起点)) % 9 + 1
+        卦表 = {1:"坎",2:"坤",3:"震",4:"巽",5:"中",6:"乾",7:"兑",8:"艮",9:"离"}
+        卦 = 卦表[飞腾数]
+        方位, 吉凶 = 子午流注类.八卦信息(卦)
+        # 配穴
+        配穴信息 = 子午流注类.飞腾配穴表.get(飞腾数, (None, None))
+        穴位, 经络 = 配穴信息 if 配穴信息[0] else ("无", "")
+        穴位详情 = 子午流注类.穴位详情(穴位)
+
+        升降表 = {
+            "坎":"宜降","坤":"宜补","震":"宜升","巽":"宜疏散",
+            "乾":"大补","兑":"慎泻","艮":"宜止","离":"宜升散","中":"平补平泻"
+        }
+        升降 = 升降表[卦]
+
+        描述 = 子午流注类.生成描述(农历)
+        摘要 = f"飞腾数：{飞腾数}（{卦}卦，{方位}，{吉凶}），{升降}，配穴：{穴位}({经络})"
+
+        return {
+            "state": 穴位,
+            "attributes": {
+                "飞腾数": 飞腾数,
+                "卦": 卦,
+                "方位": 方位,
+                "吉凶": 吉凶,
+                "升降补泻": 升降,
+                "配穴": 穴位,
+                "配穴经络": 经络,
+                "位置": 穴位详情['位置'],
+                "功能": 穴位详情['功能'],
+                "针法": 穴位详情['针法'],
+                "起点时辰": 起点,
+                "当前时辰": 时支,
+                "摘要": 摘要,
+                "描述": 描述,
+                "详细描述": (
+                    f"{描述}，飞腾八法起点为{起点}时，顺数至{时支}时得数：{飞腾数}。\n"
+                    f"对应卦象：{卦}卦（方位：{方位}，吉凶：{吉凶}）。\n"
+                    f"补泻建议：{升降}。\n"
+                    f"配穴：{穴位}（{经络}），位置：{穴位详情['位置']}，功能：{穴位详情['功能']}，针法：{穴位详情['针法']}"
+                )
+            }
+        }
+
+    # ========= 迎随补泻法 =========
+    @staticmethod
+    def 迎随补泻类(农历, dt):
+        """子午流注类 · 迎随补泻法"""
+
+        小时 = dt.hour
+        时支 = 子午流注类.小时转时支(小时)
+
+        当前经络 = 子午流注类.时辰经络[时支][0]
+        顺经 = 子午流注类.顺经(当前经络)
+        逆经 = 子午流注类.逆经(当前经络)
+
+        补法 = f"顺气补：顺着气血流注方向 → {当前经络} → {顺经}"
+        泻法 = f"逆气泻：逆着气血流注方向 → {当前经络} → {逆经}"
+
+        描述 = 子午流注类.生成描述(农历)
+        摘要 = f"{当前经络}当令，顺气补（{顺经}），逆气泻（{逆经}）"
+
+        return {
+            "state": 当前经络,
+            "attributes": {
+                "时支": 时支,
+                "当前经络": 当前经络,
+                "顺气经络": 顺经,
+                "逆气经络": 逆经,
+                "补法": 补法,
+                "泻法": 泻法,
+                "摘要": 摘要,
+                "描述": 描述,
+                "详细描述": (
+                    f"{描述}，气血流注于{当前经络}。\n"
+                    f"顺气补法：{补法}\n"
+                    f"逆气泻法：{泻法}\n"
+                    f"说明：迎随补泻法以气血流注方向为依据，顺为补，逆为泻。"
+                )
+            }
+        }
