--- custom_components/tianyuan_calendar/tianyuan/shushu.py
@@ -1,469 +1,0 @@
-"""天元术数引擎."""
-# 干支纪日：晚子时日柱算当天
-# 干支纪年：新年以立春节气交接的时刻起算
-from __future__ import annotations
-
-from typing import Any
-from .maps_loader import 加载全量映射表类, 获取单卦基础信息类, 获取所有卦名列表类, 获取单卦详注类
-
-maps = 加载全量映射表类()
-术数映射 = maps.get("术数类", {})
-
-class 天元术数类:
-
-    八卦基础属性 = 术数映射.get("八卦基础属性", {})
-    爻位 = {tuple(v["爻"]): k for k, v in 八卦基础属性.items()}
-
-# 小六壬(TianYuanShuShu)
-class 小六壬类:
-
-    小六壬映射表 = 术数映射.get("小六壬映射表", {})
-    小六壬掌诀 = ["大安", "留连", "速喜", "赤口", "小吉", "空亡"]   
-
-    @staticmethod
-    def 起卦类(农历):
-        """小六壬起卦：月上起日，日上起时"""
-        月数 = 农历.getMonth()
-        日数 = 农历.getDay()
-        时数 = 农历.getTimeZhiIndex() + 1  # 子=1
-        # 第一步：从大安起正月
-        月位 = (月数 - 1) % 6
-        # 第二步：从月位起初一
-        日位 = (月位 + 日数 - 1) % 6
-        # 第三步：从日位起子时
-        结果索引 = (日位 + 时数 - 1) % 6
-        卦名 = 小六壬类.小六壬掌诀[结果索引]
-        信息 = 小六壬类.小六壬映射表.get(卦名, {})
-
-        return {
-            "state": 卦名,
-            "attributes": {
-                "吉凶": 信息.get("吉凶", "未知"),
-                "五行": 信息.get("五行", "未知"),
-                "方位": 信息.get("方位", "未知"),
-                "用神": 信息.get("用神", "未知"),
-                "歌诀": 信息.get("歌诀", "暂无"),
-                "描述": 信息.get("描述", "暂无"),
-                "宜忌": 信息.get("宜忌", "暂无"),
-                "推算过程": f"农历 {月数}月 {日数}日 {农历.getTimeZhi()}时 → {卦名}",
-                "提醒": "马前课仅供即时参考"
-            }
-        }
-
-# 梅花易数(TianYuanShuShu)
-class 梅花易数类:
-    """梅花易数：年月日时起卦"""
-
-    @staticmethod
-    def 体用生克断语(体五行: str, 用五行: str) -> str:
-        """梅花易数体用生克断语"""
-
-        五行关系 = {
-            "金": {"生": "水", "克": "木"},
-            "木": {"生": "火", "克": "土"},
-            "水": {"生": "木", "克": "火"},
-            "火": {"生": "土", "克": "金"},
-            "土": {"生": "金", "克": "水"},
-        }
-        # 比和
-        if 体五行 == 用五行:
-            return "【比和】体用相同，谋事易成，多得助力，百祸消散。"
-        # 用生体
-        if 五行关系[用五行]["生"] == 体五行:
-            return "【用生体】主有进益之喜，贵人提拔，诸事大吉。"
-        # 体生用
-        if 五行关系[体五行]["生"] == 用五行:
-            return "【体生用】主有损耗、泄气，心力交瘁，钱财散失。"
-        # 用克体
-        if 五行关系[用五行]["克"] == 体五行:
-            return "【用克体】主有压力、灾殃，诸事受阻，需防小人。"
-        # 体克用
-        if 五行关系[体五行]["克"] == 用五行:
-            return "【体克用】主事虽有阻碍，但终能凭借自身努力克服，成事较累。"
-
-        return "体用关联观察中。"
-
-    @staticmethod
-    def 起卦类(农历):
-        """梅花易数：年月日时起卦"""
-
-        # 获取年、月、日、时（全部用地支序数）
-        年序 = 农历.getYearZhiIndex() + 1
-        月序 = 农历.getMonth()
-        日序 = 农历.getDay()
-        时序 = 农历.getTimeZhiIndex() + 1
-        # 上卦、下卦、动爻
-        上卦序 = (年序 + 月序 + 日序) % 8 or 8
-        下卦序 = (年序 + 月序 + 日序 + 时序) % 8 or 8
-        动爻序 = (年序 + 月序 + 日序 + 时序) % 6
-        # 卦象坐标
-        坐标 = f"{上卦序}-{下卦序}"
-        卦信息 = 获取单卦基础信息类(坐标)
-        # 体用判断
-        if 动爻序 <= 3:
-            体卦序, 用卦序 = 上卦序, 下卦序
-            体位, 用位 = "上卦", "下卦"
-        else:
-            体卦序, 用卦序 = 下卦序, 上卦序
-            体位, 用位 = "下卦", "上卦"
-
-        体卦 = 天元术数类.八卦基础属性[体卦序]
-        用卦 = 天元术数类.八卦基础属性[用卦序]
-        # 体用生克断语
-        断语 = 梅花易数类.体用生克断语(体卦["五行"], 用卦["五行"])
-        # 返回结构（保持你原来的格式）
-        return {
-            "state": 卦信息.get("name", "未知"),
-            "attributes": {
-                "卦辞": 卦信息.get("text", "研读中..."),
-                "象曰": 卦信息.get("image", "观察中..."),
-                "内涵": 卦信息.get("desc", "暂无描述"),
-
-                "体用断语": 断语,
-                "体卦": f"{体卦['名称']}({体卦['五行']}) - {体位}",
-                "用卦": f"{用卦['名称']}({用卦['五行']}) - {用位}",
-
-                "动爻": f"第{动爻序}爻",
-                "起卦源": f"年{年序} 月{月序} 日{日序} 时{时序}",
-            }
-        }
-
-# 皇极经世
-class 皇极经世类:
-    """皇极经世：值年卦 + 值月卦推演"""
-
-    卦名到坐标 = 术数映射.get("卦名到坐标", {})
-    年份到卦名 = 术数映射.get("年份到卦名", {})
-
-    @staticmethod
-    def 计算元会运世(公历年: int):
-        """
-        元会运世推算（邵雍体系：公元前 67017 年起算）
-        """
-        起算总年数 = 公历年 + 67017
-        已过年数 = 起算总年数 - 1
-
-        元 = 已过年数 // 129600 + 1
-        余数 = 已过年数 % 129600
-
-        会 = 余数 // 10800 + 1
-        余数 %= 10800
-
-        运 = 余数 // 360 + 1
-        余数 %= 360
-
-        世 = 余数 // 30 + 1
-        年序 = 余数 % 30 + 1
-
-        return 元, 会, 运, 世, 年序
-
-    @staticmethod
-    def 起卦类(农历, 阳历):
-        """皇极经世核心逻辑"""
-
-        # 获取基础信息
-        公历年数 = 阳历.getYear()
-        农历月数 = 农历.getMonth()
-        农历年干 = 农历.getYearGan()
-        农历年月 = 农历.getMonthInChinese()
-
-        # 年干索引：0甲 1乙 2丙 3丁 4戊 5己 6庚 7辛 8壬 9癸
-        # 偶数索引为阳年：甲、丙、戊、庚、壬
-        是否阳年 = (农历.getYearGanIndex() % 2 == 0)
-
-        # 计算元会运世
-        元, 会, 运, 世, 年序 = 皇极经世类.计算元会运世(公历年数)
-
-        # 获取值年卦
-        年卦名 = 皇极经世类.年份到卦名.get(公历年数, "未知")
-        年卦坐标 = 皇极经世类.卦名到坐标.get(年卦名, "1-1")
-        年卦信息 = 获取单卦基础信息类(年卦坐标)
-        上卦序, 下卦序 = map(int, 年卦坐标.split("-"))
-
-        # 组合六爻（下三爻 + 上三爻）
-        六爻列表 = (
-            天元术数类.八卦基础属性[下卦序]["爻"]
-            + 天元术数类.八卦基础属性[上卦序]["爻"]
-        )
-        # 推算值月卦（逐月动爻法
-        # 动爻位置（1~6）
-        动爻位置 = ((农历月数 - 1) % 6) + 1
-        if not 是否阳年:
-            # 阴年：从上往下变（6→1）
-            动爻位置 = 7 - 动爻位置
-        # 翻转爻象（1→0，0→1）
-        目标索引 = 动爻位置 - 1
-        六爻列表[目标索引] = 1 - 六爻列表[目标索引]
-        # 重新组合新卦
-        新下卦 = 天元术数类.爻位[tuple(六爻列表[:3])]
-        新上卦 = 天元术数类.爻位[tuple(六爻列表[3:])]
-        月卦坐标 = f"{新上卦}-{新下卦}"
-        月卦信息 = 获取单卦基础信息类(月卦坐标)
-        月卦名 = 月卦信息.get("name", "未知月卦")
-
-        # 返回结构
-        return {
-            "state": 月卦名,
-            "attributes": {
-                "月卦辞": 月卦信息.get("text"),
-                "月象曰": 月卦信息.get("image"),
-                "值年卦": 年卦名,
-                "年卦辞": 年卦信息.get("text"),
-                "年内涵": 年卦信息.get("desc"),
-                "历史跨度": f"第{元}元 · 第{会}会 · 第{运}运 · 第{世}世 · 第{年序}年",
-                "运行逻辑": (
-                    f"{农历年干}年（{'阳' if 是否阳年 else '阴'}年）"
-                    f"，农历{农历年月}，变第{动爻位置}爻"
-                ),
-                "系统": "皇极经世精准算法"
-            }
-        }
-
-
-# 六爻筮法
-class 六爻占卜类:
-    """六爻占卜引擎"""
-
-    天干五行 = 术数映射.get("六爻天干五行", {})
-    五行生克 = 术数映射.get("六爻五行生克", {})
-    纳甲表 = 术数映射.get("六爻纳甲表", {})
-    八宫世应表 = 术数映射.get("八宫世应表", {})
-    地支五行表 = 术数映射.get("地支五行", {})
-    宫卦序表 = 术数映射.get("卦序表", [])
-
-    # ========= 六亲判定 =========
-    @staticmethod
-    def 判六亲法(爻五行: str, 日干五行: str) -> str:
-        关系 = 六爻占卜类.五行生克
-
-        if 爻五行 == 日干五行:
-            return "兄弟"
-        if 关系[爻五行]["生"] == 日干五行:
-            return "父母"
-        if 关系[日干五行]["生"] == 爻五行:
-            return "子孙"
-        if 关系[爻五行]["克"] == 日干五行:
-            return "官鬼"
-        if 关系[日干五行]["克"] == 爻五行:
-            return "妻财"
-
-        return "六亲待察"
-
-    # ========= 纳甲展开 =========
-    @staticmethod
-    def 展开纳甲法(上卦序: int, 下卦序: int):
-        卦表 = 天元术数类.八卦基础属性
-        下卦名 = 卦表[下卦序]["名称"]
-        上卦名 = 卦表[上卦序]["名称"]
-
-        return (
-            六爻占卜类.纳甲表[下卦名]
-            + 六爻占卜类.纳甲表[上卦名]
-        )
-
-    @staticmethod
-    def 干支地支五行法(干支: str) -> str:
-        """取地支判定五行"""
-        if len(干支) >= 2:
-            return 六爻占卜类.地支五行表.get(干支[1], "未知")
-        return "未知"
-
-    # ========= 自动世应 =========
-    @staticmethod
-    def 计算世应法(上卦序, 下卦序):
-        键 = f"{上卦序}-{下卦序}"
-        return 六爻占卜类.八宫世应表.get(键, (None, None))
-
-# ========= 字符串解析法 =========
-    @staticmethod
-    def 执行占卜流程类(输入字符串: str, 农历):
-        """
-        全流程包装方法：解析 -> 校验 -> 占卜
-        协调器只需一行代码调用此方法。
-        """
-        # 1. 处理空输入
-        if not 输入字符串 or str(输入字符串).strip() == "":
-            return {
-                "state": "等待摇卦",
-                "attributes": {"提示": "输入格式：阴阳流-动爻。阳为1，阴为0。"}
-            }
-
-        # 2. 解析逻辑（将原解析法整合于此）
-        try:
-            # 预处理：归一化
-            处理串 = str(输入字符串).replace(" ", "").replace("阳", "1").replace("阴", "0")
-            
-            # 分割阴阳和动爻
-            if "-" in 处理串:
-                主体部分, 动部分 = 处理串.split("-", 1)
-            else:
-                主体部分, 动部分 = 处理串, ""
-
-            # 提取 6 位阴阳
-            主体数字串 = 主体部分.replace(",", "")
-            if len(主体数字串) != 6:
-                raise ValueError("长度不符")
-            
-            阴阳列表 = [int(x) for x in 主体数字串]
-
-            # 提取动爻索引 (1-6)
-            import re
-            动爻索引集 = [int(x) for x in re.findall(r'[1-6]', 动部分)]
-
-            # 构建六爻结构列表
-            六爻列表 = []
-            for i in range(6):
-                六爻列表.append({
-                    "爻": 阴阳列表[i],
-                    "moving": (i + 1) in 动爻索引集
-                })
-
-            # 3. 校验通过，调用原有的自动占卜逻辑类
-            # 注意：此处确保调用的是你已经写好的 自动占卜类 方法
-            return 六爻占卜类.自动占卜类(六爻列表, 农历)
-
-        except Exception:
-            # 4. 解析失败，返回错误字典，协调器依然直接透传
-            return {
-                "state": "格式错误",
-                "summary": "输入格式不合法",
-                "attributes": {
-                    "当前输入": 输入字符串,
-                    "正确示例": "101011-25 或 阳阴阳阳阴阳-6"
-                }
-            }
-        
-    # ========= 自动入口 =========
-    @staticmethod
-    def 自动占卜类(六爻列表, 农历):
-        # 自动日干
-        日干 = 农历.getDayGanExact2()
-
-        # 构建本卦爻值
-        爻值 = [1 if l["爻"] else 0 for l in 六爻列表]
-        下卦 = tuple(爻值[0:3])
-        上卦 = tuple(爻值[3:6])
-
-        # 获取卦序（依赖天元术数类，若缺失需补全）
-        上卦序 = 天元术数类.爻位[上卦]
-        下卦序 = 天元术数类.爻位[下卦]
-
-        # 修正：同时传入上下卦序计算世应
-        世位, 应位 = 六爻占卜类.计算世应法(上卦序, 下卦序)
-
-        return 六爻占卜类.占卜类(
-            六爻列表=六爻列表,
-            日干=日干,
-            世位=世位,
-            应位=应位
-        )
-
-    # ========= 主计算函数 =========
-    @staticmethod
-    def 占卜类(六爻列表, 日干, 世位=None, 应位=None):
-        if len(六爻列表) != 6:
-            raise ValueError("六爻必须是6条，自下而上")
-
-        # 1. 本卦
-        爻值 = [1 if l["爻"] else 0 for l in 六爻列表]
-        上卦 = tuple(爻值[3:6])
-        下卦 = tuple(爻值[0:3])
-        上卦序 = 天元术数类.爻位[上卦]
-        下卦序 = 天元术数类.爻位[下卦]
-        坐标 = f"{上卦序}-{下卦序}"
-
-        本卦信息 = 获取单卦基础信息类(坐标)
-
-        # 2. 变卦
-        变爻值 = [
-            (0 if l["爻"] == 1 else 1) if l.get("moving") else l["爻"]
-            for l in 六爻列表
-        ]
-        上卦变 = tuple(变爻值[3:6])
-        下卦变 = tuple(变爻值[0:3])
-        上卦变序 = 天元术数类.爻位[上卦变]
-        下卦变序 = 天元术数类.爻位[下卦变]
-        坐标变 = f"{上卦变序}-{下卦变序}"
-        变卦信息 = 获取单卦基础信息类(坐标变)
-
-        # 3. 纳甲（本卦）
-        纳甲列表 = 六爻占卜类.展开纳甲法(上卦序, 下卦序)
-
-        # 4. 卦宫五行（传统六亲依据）
-        宫卦序 = 六爻占卜类.宫卦序表.get(坐标, 0)
-        宫卦五行 = 天元术数类.八卦基础属性[宫卦序]["五行"] if 宫卦序 else "未知"
-
-        # 5. 六爻详情
-        六爻详情 = []
-        for i in range(6):
-            爻位 = i + 1
-            干支 = 纳甲列表[i]
-            爻五行 = 六爻占卜类.干支地支五行法(干支)    # 用地支五行
-            六亲 = 六爻占卜类.判六亲法(爻五行, 宫卦五行)
-            六爻详情.append({
-                "爻位": 爻位,
-                "阴阳": "阳爻" if 六爻列表[i]["爻"] == 1 else "阴爻",
-                "动静": "动爻" if 六爻列表[i].get("moving") else "静爻",
-                "纳甲": 干支,
-                "五行": 爻五行,
-                "六亲": 六亲,
-                "世": (爻位 == 世位),
-                "应": (爻位 == 应位),
-            })
-
-        爻位名 = ["初爻", "二爻", "三爻", "四爻", "五爻", "六爻"]
-        独立爻属性 = {}
-        for 爻 in 六爻详情:
-            键 = 爻位名[爻["爻位"] - 1]   # 爻位从1开始
-            值 = (
-                f"阴阳: {爻['阴阳']}, 动静: {爻['动静']}, "
-                f"纳甲: {爻['纳甲']}, 五行: {爻['五行']}, "
-                f"六亲: {爻['六亲']}, 世: {爻['世']}, 应: {爻['应']}"
-            )
-            独立爻属性[键] = 值
-
-        return {
-            "state": 本卦信息.get("name", "未知本卦"),
-            "attributes": {
-                "本卦名称": 本卦信息.get("name", ""),
-                "本卦卦辞": 本卦信息.get("text", ""),
-                "本卦象曰": 本卦信息.get("image", ""),
-                "本卦内涵": 本卦信息.get("desc", ""),
-                "变卦名称": 变卦信息.get("name", ""),
-                "变卦卦辞": 变卦信息.get("text", ""),
-                "变卦象曰": 变卦信息.get("image", ""),
-                "变卦内涵": 变卦信息.get("desc", ""),
-                **独立爻属性,
-                "日干": 日干,
-                "日干五行": 六爻占卜类.天干五行.get(日干, "未知"),
-                "宫卦五行": 宫卦五行
-            },
-        }
-
-# 易经详注库类
-class 易经详注类:
-    """负责单卦的高级详注提取."""
-
-    @staticmethod
-    def 获取易经卦象所有卦名类() -> list[str]:
-        """获取所有卦名（供 Select 实体选项使用）"""
-        return 获取所有卦名列表类()
-
-    @classmethod
-    def 获取易经卦象类(cls, 卦名: str) -> dict[str, Any]:
-        """按卦名单独提取原始详注字典"""
-        # 增加基础容错：如果卦名为 None 或 "未知"，返回空字典
-        if not 卦名 or 卦名 == "未知":
-            return {}
-        return 获取单卦详注类(卦名)
-
-    @classmethod
-    def 获取详注包装类(cls, 卦名: str) -> dict[str, Any]:
-
-        显示卦名 = 卦名 if 卦名 and 卦名 != "未知" else "乾为天"
-        卦详细资料 = cls.获取易经卦象类(显示卦名)
-        
-        return {
-            "state": 显示卦名,
-            "attributes": 卦详细资料  
-        }
