+++ custom_components/tianyuan_calendar/tianyuan/shushu/__init__.py
@@ -1,0 +1,15 @@
+"""天元术数引擎."""
+from .base import 天元术数类
+from .xiaoliuren import 小六壬类
+from .meihua import 梅花易数类
+from .huangji import 皇极经世类
+from .liuyao import 六爻占卜类
+from .xiangzhu import 易经详注类
+
+__all__ = [
+    "小六壬类",
+    "梅花易数类",
+    "皇极经世类",
+    "六爻占卜类",
+    "易经详注类",
+]
