+++ custom_components/tianyuan_calendar/tianyuan/shushu/huangji.py
@@ -1,0 +1,98 @@
+"""天元术数引擎 - 皇极经世."""
+from __future__ import annotations
+
+from ..library import 加载全量映射表类, 获取单卦基础信息类
+from .base import 天元术数类
+
+maps = 加载全量映射表类()
+术数映射 = maps.get("术数类", {})
+
+class 皇极经世类:
+    """皇极经世：值年卦 + 值月卦推演"""
+
+    卦名到坐标 = 术数映射.get("卦名到坐标", {})
+    年份到卦名 = 术数映射.get("年份到卦名", {})
+
+    @staticmethod
+    def 计算元会运世(公历年: int):
+        """
+        元会运世推算（邵雍体系：公元前 67017 年起算）
+        """
+        起算总年数 = 公历年 + 67017
+        已过年数 = 起算总年数 - 1
+
+        元 = 已过年数 // 129600 + 1
+        余数 = 已过年数 % 129600
+
+        会 = 余数 // 10800 + 1
+        余数 %= 10800
+
+        运 = 余数 // 360 + 1
+        余数 %= 360
+
+        世 = 余数 // 30 + 1
+        年序 = 余数 % 30 + 1
+
+        return 元, 会, 运, 世, 年序
+
+    @staticmethod
+    def 起卦类(农历, 阳历):
+        """皇极经世核心逻辑"""
+
+        # 获取基础信息
+        公历年数 = 阳历.getYear()
+        农历月数 = 农历.getMonth()
+        农历年干 = 农历.getYearGan()
+        农历年月 = 农历.getMonthInChinese()
+
+        # 年干索引：0甲 1乙 2丙 3丁 4戊 5己 6庚 7辛 8壬 9癸
+        # 偶数索引为阳年：甲、丙、戊、庚、壬
+        是否阳年 = (农历.getYearGanIndex() % 2 == 0)
+
+        # 计算元会运世
+        元, 会, 运, 世, 年序 = 皇极经世类.计算元会运世(公历年数)
+
+        # 获取值年卦
+        年卦名 = 皇极经世类.年份到卦名.get(公历年数, "未知")
+        年卦坐标 = 皇极经世类.卦名到坐标.get(年卦名, "1-1")
+        年卦信息 = 获取单卦基础信息类(年卦坐标)
+        上卦序, 下卦序 = map(int, 年卦坐标.split("-"))
+
+        # 组合六爻（下三爻 + 上三爻）
+        六爻列表 = (
+            天元术数类.八卦基础属性[下卦序]["爻"]
+            + 天元术数类.八卦基础属性[上卦序]["爻"]
+        )
+        # 推算值月卦（逐月动爻法
+        # 动爻位置（1~6）
+        动爻位置 = ((农历月数 - 1) % 6) + 1
+        if not 是否阳年:
+            # 阴年：从上往下变（6→1）
+            动爻位置 = 7 - 动爻位置
+        # 翻转爻象（1→0，0→1）
+        目标索引 = 动爻位置 - 1
+        六爻列表[目标索引] = 1 - 六爻列表[目标索引]
+        # 重新组合新卦
+        新下卦 = 天元术数类.爻位[tuple(六爻列表[:3])]
+        新上卦 = 天元术数类.爻位[tuple(六爻列表[3:])]
+        月卦坐标 = f"{新上卦}-{新下卦}"
+        月卦信息 = 获取单卦基础信息类(月卦坐标)
+        月卦名 = 月卦信息.get("name", "未知月卦")
+
+        # 返回结构
+        return {
+            "state": 月卦名,
+            "attributes": {
+                "月卦辞": 月卦信息.get("text"),
+                "月象曰": 月卦信息.get("image"),
+                "值年卦": 年卦名,
+                "年卦辞": 年卦信息.get("text"),
+                "年内涵": 年卦信息.get("desc"),
+                "历史跨度": f"第{元}元 · 第{会}会 · 第{运}运 · 第{世}世 · 第{年序}年",
+                "运行逻辑": (
+                    f"{农历年干}年（{'阳' if 是否阳年 else '阴'}年）"
+                    f"，农历{农历年月}，变第{动爻位置}爻"
+                ),
+                "系统": "皇极经世精准算法"
+            }
+        }
