+++ custom_components/tianyuan_calendar/tianyuan/shushu/liuyao.py
@@ -1,0 +1,227 @@
+"""天元术数引擎 - 六爻占卜."""
+from __future__ import annotations
+
+from ..library import 加载全量映射表类, 获取单卦基础信息类
+from .base import 天元术数类
+
+maps = 加载全量映射表类()
+术数映射 = maps.get("术数类", {})
+
+class 六爻占卜类:
+    """六爻占卜引擎"""
+
+    天干五行 = 术数映射.get("六爻天干五行", {})
+    五行生克 = 术数映射.get("六爻五行生克", {})
+    纳甲表 = 术数映射.get("六爻纳甲表", {})
+    八宫世应表 = 术数映射.get("八宫世应表", {})
+    地支五行表 = 术数映射.get("地支五行", {})
+    宫卦序表 = 术数映射.get("卦序表", [])
+
+    # ========= 六亲判定 =========
+    @staticmethod
+    def 判六亲法(爻五行: str, 日干五行: str) -> str:
+        关系 = 六爻占卜类.五行生克
+
+        if 爻五行 == 日干五行:
+            return "兄弟"
+        if 关系[爻五行]["生"] == 日干五行:
+            return "父母"
+        if 关系[日干五行]["生"] == 爻五行:
+            return "子孙"
+        if 关系[爻五行]["克"] == 日干五行:
+            return "官鬼"
+        if 关系[日干五行]["克"] == 爻五行:
+            return "妻财"
+
+        return "六亲待察"
+
+    # ========= 纳甲展开 =========
+    @staticmethod
+    def 展开纳甲法(上卦序: int, 下卦序: int):
+        卦表 = 天元术数类.八卦基础属性
+        下卦名 = 卦表[下卦序]["名称"]
+        上卦名 = 卦表[上卦序]["名称"]
+
+        return (
+            六爻占卜类.纳甲表[下卦名]
+            + 六爻占卜类.纳甲表[上卦名]
+        )
+
+    @staticmethod
+    def 干支地支五行法(干支: str) -> str:
+        """取地支判定五行"""
+        if len(干支) >= 2:
+            return 六爻占卜类.地支五行表.get(干支[1], "未知")
+        return "未知"
+
+    # ========= 自动世应 =========
+    @staticmethod
+    def 计算世应法(上卦序, 下卦序):
+        键 = f"{上卦序}-{下卦序}"
+        return 六爻占卜类.八宫世应表.get(键, (None, None))
+
+# ========= 字符串解析法 =========
+    @staticmethod
+    def 执行占卜流程类(输入字符串: str, 农历):
+        """
+        全流程包装方法：解析 -> 校验 -> 占卜
+        协调器只需一行代码调用此方法。
+        """
+        # 1. 处理空输入
+        if not 输入字符串 or str(输入字符串).strip() == "":
+            return {
+                "state": "等待摇卦",
+                "attributes": {"提示": "输入格式：阴阳流-动爻。阳为1，阴为0。"}
+            }
+
+        # 2. 解析逻辑（将原解析法整合于此）
+        try:
+            # 预处理：归一化
+            处理串 = str(输入字符串).replace(" ", "").replace("阳", "1").replace("阴", "0")
+            
+            # 分割阴阳和动爻
+            if "-" in 处理串:
+                主体部分, 动部分 = 处理串.split("-", 1)
+            else:
+                主体部分, 动部分 = 处理串, ""
+
+            # 提取 6 位阴阳
+            主体数字串 = 主体部分.replace(",", "")
+            if len(主体数字串) != 6:
+                raise ValueError("长度不符")
+            
+            阴阳列表 = [int(x) for x in 主体数字串]
+
+            # 提取动爻索引 (1-6)
+            import re
+            动爻索引集 = [int(x) for x in re.findall(r'[1-6]', 动部分)]
+
+            # 构建六爻结构列表
+            六爻列表 = []
+            for i in range(6):
+                六爻列表.append({
+                    "爻": 阴阳列表[i],
+                    "moving": (i + 1) in 动爻索引集
+                })
+
+            # 3. 校验通过，调用原有的自动占卜逻辑类
+            # 注意：此处确保调用的是你已经写好的 自动占卜类 方法
+            return 六爻占卜类.自动占卜类(六爻列表, 农历)
+
+        except Exception:
+            # 4. 解析失败，返回错误字典，协调器依然直接透传
+            return {
+                "state": "格式错误",
+                "summary": "输入格式不合法",
+                "attributes": {
+                    "当前输入": 输入字符串,
+                    "正确示例": "101011-25 或 阳阴阳阳阴阳-6"
+                }
+            }
+        
+    # ========= 自动入口 =========
+    @staticmethod
+    def 自动占卜类(六爻列表, 农历):
+        # 自动日干
+        日干 = 农历.getDayGanExact2()
+
+        # 构建本卦爻值
+        爻值 = [1 if l["爻"] else 0 for l in 六爻列表]
+        下卦 = tuple(爻值[0:3])
+        上卦 = tuple(爻值[3:6])
+
+        # 获取卦序（依赖天元术数类，若缺失需补全）
+        上卦序 = 天元术数类.爻位[上卦]
+        下卦序 = 天元术数类.爻位[下卦]
+
+        # 修正：同时传入上下卦序计算世应
+        世位, 应位 = 六爻占卜类.计算世应法(上卦序, 下卦序)
+
+        return 六爻占卜类.占卜类(
+            六爻列表=六爻列表,
+            日干=日干,
+            世位=世位,
+            应位=应位
+        )
+
+    # ========= 主计算函数 =========
+    @staticmethod
+    def 占卜类(六爻列表, 日干, 世位=None, 应位=None):
+        if len(六爻列表) != 6:
+            raise ValueError("六爻必须是6条，自下而上")
+
+        # 1. 本卦
+        爻值 = [1 if l["爻"] else 0 for l in 六爻列表]
+        上卦 = tuple(爻值[3:6])
+        下卦 = tuple(爻值[0:3])
+        上卦序 = 天元术数类.爻位[上卦]
+        下卦序 = 天元术数类.爻位[下卦]
+        坐标 = f"{上卦序}-{下卦序}"
+
+        本卦信息 = 获取单卦基础信息类(坐标)
+
+        # 2. 变卦
+        变爻值 = [
+            (0 if l["爻"] == 1 else 1) if l.get("moving") else l["爻"]
+            for l in 六爻列表
+        ]
+        上卦变 = tuple(变爻值[3:6])
+        下卦变 = tuple(变爻值[0:3])
+        上卦变序 = 天元术数类.爻位[上卦变]
+        下卦变序 = 天元术数类.爻位[下卦变]
+        坐标变 = f"{上卦变序}-{下卦变序}"
+        变卦信息 = 获取单卦基础信息类(坐标变)
+
+        # 3. 纳甲（本卦）
+        纳甲列表 = 六爻占卜类.展开纳甲法(上卦序, 下卦序)
+
+        # 4. 卦宫五行（传统六亲依据）
+        宫卦序 = 六爻占卜类.宫卦序表.get(坐标, 0)
+        宫卦五行 = 天元术数类.八卦基础属性[宫卦序]["五行"] if 宫卦序 else "未知"
+
+        # 5. 六爻详情
+        六爻详情 = []
+        for i in range(6):
+            爻位 = i + 1
+            干支 = 纳甲列表[i]
+            爻五行 = 六爻占卜类.干支地支五行法(干支)    # 用地支五行
+            六亲 = 六爻占卜类.判六亲法(爻五行, 宫卦五行)
+            六爻详情.append({
+                "爻位": 爻位,
+                "阴阳": "阳爻" if 六爻列表[i]["爻"] == 1 else "阴爻",
+                "动静": "动爻" if 六爻列表[i].get("moving") else "静爻",
+                "纳甲": 干支,
+                "五行": 爻五行,
+                "六亲": 六亲,
+                "世": (爻位 == 世位),
+                "应": (爻位 == 应位),
+            })
+
+        爻位名 = ["初爻", "二爻", "三爻", "四爻", "五爻", "六爻"]
+        独立爻属性 = {}
+        for 爻 in 六爻详情:
+            键 = 爻位名[爻["爻位"] - 1]   # 爻位从1开始
+            值 = (
+                f"阴阳: {爻['阴阳']}, 动静: {爻['动静']}, "
+                f"纳甲: {爻['纳甲']}, 五行: {爻['五行']}, "
+                f"六亲: {爻['六亲']}, 世: {爻['世']}, 应: {爻['应']}"
+            )
+            独立爻属性[键] = 值
+
+        return {
+            "state": 本卦信息.get("name", "未知本卦"),
+            "attributes": {
+                "本卦名称": 本卦信息.get("name", ""),
+                "本卦卦辞": 本卦信息.get("text", ""),
+                "本卦象曰": 本卦信息.get("image", ""),
+                "本卦内涵": 本卦信息.get("desc", ""),
+                "变卦名称": 变卦信息.get("name", ""),
+                "变卦卦辞": 变卦信息.get("text", ""),
+                "变卦象曰": 变卦信息.get("image", ""),
+                "变卦内涵": 变卦信息.get("desc", ""),
+                **独立爻属性,
+                "日干": 日干,
+                "日干五行": 六爻占卜类.天干五行.get(日干, "未知"),
+                "宫卦五行": 宫卦五行
+            },
+        }
