+++ custom_components/tianyuan_calendar/tianyuan/shushu/meihua.py
@@ -1,0 +1,82 @@
+"""天元术数引擎 - 梅花易数."""
+from __future__ import annotations
+
+from ..library import 获取单卦基础信息类
+from .base import 天元术数类
+
+class 梅花易数类:
+    """梅花易数：年月日时起卦"""
+
+    @staticmethod
+    def 体用生克断语(体五行: str, 用五行: str) -> str:
+        """梅花易数体用生克断语"""
+
+        五行关系 = {
+            "金": {"生": "水", "克": "木"},
+            "木": {"生": "火", "克": "土"},
+            "水": {"生": "木", "克": "火"},
+            "火": {"生": "土", "克": "金"},
+            "土": {"生": "金", "克": "水"},
+        }
+        # 比和
+        if 体五行 == 用五行:
+            return "【比和】体用相同，谋事易成，多得助力，百祸消散。"
+        # 用生体
+        if 五行关系[用五行]["生"] == 体五行:
+            return "【用生体】主有进益之喜，贵人提拔，诸事大吉。"
+        # 体生用
+        if 五行关系[体五行]["生"] == 用五行:
+            return "【体生用】主有损耗、泄气，心力交瘁，钱财散失。"
+        # 用克体
+        if 五行关系[用五行]["克"] == 体五行:
+            return "【用克体】主有压力、灾殃，诸事受阻，需防小人。"
+        # 体克用
+        if 五行关系[体五行]["克"] == 用五行:
+            return "【体克用】主事虽有阻碍，但终能凭借自身努力克服，成事较累。"
+
+        return "体用关联观察中。"
+
+    @staticmethod
+    def 起卦类(农历):
+        """梅花易数：年月日时起卦"""
+
+        # 获取年、月、日、时（全部用地支序数）
+        年序 = 农历.getYearZhiIndex() + 1
+        月序 = 农历.getMonth()
+        日序 = 农历.getDay()
+        时序 = 农历.getTimeZhiIndex() + 1
+        # 上卦、下卦、动爻
+        上卦序 = (年序 + 月序 + 日序) % 8 or 8
+        下卦序 = (年序 + 月序 + 日序 + 时序) % 8 or 8
+        动爻序 = (年序 + 月序 + 日序 + 时序) % 6
+        # 卦象坐标
+        坐标 = f"{上卦序}-{下卦序}"
+        卦信息 = 获取单卦基础信息类(坐标)
+        # 体用判断
+        if 动爻序 <= 3:
+            体卦序, 用卦序 = 上卦序, 下卦序
+            体位, 用位 = "上卦", "下卦"
+        else:
+            体卦序, 用卦序 = 下卦序, 上卦序
+            体位, 用位 = "下卦", "上卦"
+
+        体卦 = 天元术数类.八卦基础属性[体卦序]
+        用卦 = 天元术数类.八卦基础属性[用卦序]
+        # 体用生克断语
+        断语 = 梅花易数类.体用生克断语(体卦["五行"], 用卦["五行"])
+        # 返回结构（保持你原来的格式）
+        return {
+            "state": 卦信息.get("name", "未知"),
+            "attributes": {
+                "卦辞": 卦信息.get("text", "研读中..."),
+                "象曰": 卦信息.get("image", "观察中..."),
+                "内涵": 卦信息.get("desc", "暂无描述"),
+
+                "体用断语": 断语,
+                "体卦": f"{体卦['名称']}({体卦['五行']}) - {体位}",
+                "用卦": f"{用卦['名称']}({用卦['五行']}) - {用位}",
+
+                "动爻": f"第{动爻序}爻",
+                "起卦源": f"年{年序} 月{月序} 日{日序} 时{时序}",
+            }
+        }
