+++ custom_components/tianyuan_calendar/tianyuan/shushu/xiaoliuren.py
@@ -1,0 +1,42 @@
+"""天元术数引擎 - 小六壬."""
+from __future__ import annotations
+
+from ..library import 加载全量映射表类
+
+maps = 加载全量映射表类()
+术数映射 = maps.get("术数类", {})
+
+class 小六壬类:
+
+    小六壬映射表 = 术数映射.get("小六壬映射表", {})
+    小六壬掌诀 = ["大安", "留连", "速喜", "赤口", "小吉", "空亡"]   
+
+    @staticmethod
+    def 起卦类(农历):
+        """小六壬起卦：月上起日，日上起时"""
+        月数 = 农历.getMonth()
+        日数 = 农历.getDay()
+        时数 = 农历.getTimeZhiIndex() + 1  # 子=1
+        # 第一步：从大安起正月
+        月位 = (月数 - 1) % 6
+        # 第二步：从月位起初一
+        日位 = (月位 + 日数 - 1) % 6
+        # 第三步：从日位起子时
+        结果索引 = (日位 + 时数 - 1) % 6
+        卦名 = 小六壬类.小六壬掌诀[结果索引]
+        信息 = 小六壬类.小六壬映射表.get(卦名, {})
+
+        return {
+            "state": 卦名,
+            "attributes": {
+                "吉凶": 信息.get("吉凶", "未知"),
+                "五行": 信息.get("五行", "未知"),
+                "方位": 信息.get("方位", "未知"),
+                "用神": 信息.get("用神", "未知"),
+                "歌诀": 信息.get("歌诀", "暂无"),
+                "描述": 信息.get("描述", "暂无"),
+                "宜忌": 信息.get("宜忌", "暂无"),
+                "推算过程": f"农历 {月数}月 {日数}日 {农历.getTimeZhi()}时 → {卦名}",
+                "提醒": "马前课仅供即时参考"
+            }
+        }
